@@ -1,18 +1,22 @@
 name: fl_chart
 description: A powerful Flutter chart library, currently supporting Line Chart, Bar Chart and Pie Chart.
-version: 0.9.4
+version: 0.30.0
 homepage: https://github.com/imaNNeoFighT/fl_chart
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  path_drawing: ^0.4.1
-  equatable: ^1.1.1
+  equatable: ^2.0.0
+  pedantic: ^1.10.0
+
+  flutter_svg: ^0.21.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+
