@@ -1,35 +1,32 @@
 import 'dart:math';
+import 'dart:ui' as ui;
 
 import 'package:fl_chart/fl_chart.dart';
 import 'package:fl_chart/src/chart/bar_chart/bar_chart_data.dart';
 import 'package:fl_chart/src/chart/base/axis_chart/axis_chart_painter.dart';
 import 'package:fl_chart/src/chart/base/base_chart/base_chart_painter.dart';
+import 'package:fl_chart/src/utils/canvas_wrapper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'bar_chart_extensions.dart';
 
 import '../../utils/utils.dart';
 
 /// Paints [BarChartData] in the canvas, it can be used in a [CustomPainter]
-class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<BarTouchResponse> {
-  Paint _barPaint, _bgTouchTooltipPaint;
+class BarChartPainter extends AxisChartPainter<BarChartData> {
+  late Paint _barPaint, _bgTouchTooltipPaint;
 
-  List<_GroupBarsPosition> _groupBarsPosition;
+  List<_GroupBarsPosition>? _groupBarsPosition;
 
   /// Paints [data] into canvas, it is the animating [BarChartData],
   /// [targetData] is the animation's target and remains the same
   /// during animation, then we should use it  when we need to show
   /// tooltips or something like that, because [data] is changing constantly.
   ///
-  /// [touchHandler] passes a [TouchHandler] to the parent,
-  /// parent will use it for touch handling flow.
-  ///
   /// [textScale] used for scaling texts inside the chart,
   /// parent can use [MediaQuery.textScaleFactor] to respect
   /// the system's font size.
-  BarChartPainter(BarChartData data, BarChartData targetData, Function(TouchHandler) touchHandler,
-      {double textScale = 1})
-      : super(data, targetData, textScale: textScale) {
-    touchHandler(this);
+  BarChartPainter() : super() {
     _barPaint = Paint()..style = PaintingStyle.fill;
 
     _bgTouchTooltipPaint = Paint()
@@ -39,46 +36,50 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
 
   /// Paints [BarChartData] into the provided canvas.
   @override
-  void paint(Canvas canvas, Size size) {
-    super.paint(canvas, size);
+  void paint(Canvas canvas, Size size, PaintHolder<BarChartData> holder) {
+    super.paint(canvas, size, holder);
+    final data = holder.data;
+    final targetData = holder.targetData;
+    final canvasWrapper = CanvasWrapper(canvas, size);
 
     if (data.barGroups.isEmpty) {
       return;
     }
 
-    final List<double> groupsX = _calculateGroupsX(size, data.barGroups, data.alignment);
+    final groupsX = _calculateGroupsX(size, data.barGroups, data.alignment, holder);
     _groupBarsPosition = _calculateGroupAndBarsPosition(size, groupsX, data.barGroups);
 
-    _drawBars(canvas, size, _groupBarsPosition);
-    drawAxisTitles(canvas, size);
-    _drawTitles(canvas, size, _groupBarsPosition);
+    _drawBars(canvasWrapper, _groupBarsPosition!, holder);
+    drawAxisTitles(canvasWrapper, holder);
+    _drawTitles(canvasWrapper, _groupBarsPosition!, holder);
 
-    for (int i = 0; i < targetData.barGroups.length; i++) {
+    for (var i = 0; i < targetData.barGroups.length; i++) {
       final barGroup = targetData.barGroups[i];
-      for (int j = 0; j < barGroup.barRods.length; j++) {
+      for (var j = 0; j < barGroup.barRods.length; j++) {
         if (!barGroup.showingTooltipIndicators.contains(j)) {
           continue;
         }
         final barRod = barGroup.barRods[j];
 
-        _drawTouchTooltip(canvas, size, _groupBarsPosition,
-            targetData.barTouchData.touchTooltipData, barGroup, i, barRod, j);
+        _drawTouchTooltip(canvasWrapper, _groupBarsPosition!,
+            targetData.barTouchData.touchTooltipData, barGroup, i, barRod, j, holder);
       }
     }
   }
 
   /// Calculates groups position for showing in the x axis using [alignment].
-  List<double> _calculateGroupsX(
-      Size viewSize, List<BarChartGroupData> barGroups, BarChartAlignment alignment) {
-    final Size drawSize = getChartUsableDrawSize(viewSize);
+  List<double> _calculateGroupsX(Size viewSize, List<BarChartGroupData> barGroups,
+      BarChartAlignment alignment, PaintHolder<BarChartData> holder) {
+    final data = holder.data;
+    final drawSize = getChartUsableDrawSize(viewSize, holder);
 
-    final List<double> groupsX = List<double>(barGroups.length);
+    final groupsX = List.filled(barGroups.length, 0.0, growable: false);
 
-    final double leftTextsSpace = getLeftOffsetDrawSize();
+    final leftTextsSpace = getLeftOffsetDrawSize(holder);
 
     switch (alignment) {
       case BarChartAlignment.start:
-        double tempX = 0;
+        var tempX = 0.0;
         barGroups.asMap().forEach((i, group) {
           groupsX[i] = leftTextsSpace + tempX + group.width / 2;
           tempX += group.width;
@@ -86,8 +87,8 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
         break;
 
       case BarChartAlignment.end:
-        double tempX = 0;
-        for (int i = barGroups.length - 1; i >= 0; i--) {
+        var tempX = 0.0;
+        for (var i = barGroups.length - 1; i >= 0; i--) {
           final group = barGroups[i];
           groupsX[i] = (leftTextsSpace + drawSize.width) - tempX - group.width / 2;
           tempX += group.width;
@@ -95,26 +96,26 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
         break;
 
       case BarChartAlignment.center:
-        double sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
+        var sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
         sumWidth += data.groupsSpace * (barGroups.length - 1);
-        final double horizontalMargin = (drawSize.width - sumWidth) / 2;
+        final horizontalMargin = (drawSize.width - sumWidth) / 2;
 
-        double tempX = 0;
-        for (int i = 0; i < barGroups.length; i++) {
+        var tempX = 0.0;
+        for (var i = 0; i < barGroups.length; i++) {
           final group = barGroups[i];
           groupsX[i] = leftTextsSpace + horizontalMargin + tempX + group.width / 2;
 
-          final double groupSpace = i == barGroups.length - 1 ? 0 : data.groupsSpace;
+          final groupSpace = i == barGroups.length - 1 ? 0 : data.groupsSpace;
           tempX += group.width + groupSpace;
         }
         break;
 
       case BarChartAlignment.spaceBetween:
-        final double sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final double spaceAvailable = drawSize.width - sumWidth;
-        final double eachSpace = spaceAvailable / (barGroups.length - 1);
+        final sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
+        final spaceAvailable = drawSize.width - sumWidth;
+        final eachSpace = spaceAvailable / (barGroups.length - 1);
 
-        double tempX = 0;
+        var tempX = 0.0;
         barGroups.asMap().forEach((index, group) {
           tempX += group.width / 2;
           if (index != 0) {
@@ -126,11 +127,11 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
         break;
 
       case BarChartAlignment.spaceAround:
-        final double sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final double spaceAvailable = drawSize.width - sumWidth;
-        final double eachSpace = spaceAvailable / (barGroups.length * 2);
+        final sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
+        final spaceAvailable = drawSize.width - sumWidth;
+        final eachSpace = spaceAvailable / (barGroups.length * 2);
 
-        double tempX = 0;
+        var tempX = 0.0;
         barGroups.asMap().forEach((i, group) {
           tempX += eachSpace;
           tempX += group.width / 2;
@@ -141,11 +142,11 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
         break;
 
       case BarChartAlignment.spaceEvenly:
-        final double sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
-        final double spaceAvailable = drawSize.width - sumWidth;
-        final double eachSpace = spaceAvailable / (barGroups.length + 1);
+        final sumWidth = barGroups.map((group) => group.width).reduce((a, b) => a + b);
+        final spaceAvailable = drawSize.width - sumWidth;
+        final eachSpace = spaceAvailable / (barGroups.length + 1);
 
-        double tempX = 0;
+        var tempX = 0.0;
         barGroups.asMap().forEach((i, group) {
           tempX += eachSpace;
           tempX += group.width / 2;
@@ -165,15 +166,15 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
       throw Exception('inconsistent state groupsX.length != barGroups.length');
     }
 
-    final List<_GroupBarsPosition> groupBarsPosition = [];
-    for (int i = 0; i < barGroups.length; i++) {
-      final BarChartGroupData barGroup = barGroups[i];
-      final double groupX = groupsX[i];
+    final groupBarsPosition = <_GroupBarsPosition>[];
+    for (var i = 0; i < barGroups.length; i++) {
+      final barGroup = barGroups[i];
+      final groupX = groupsX[i];
 
-      double tempX = 0;
-      final List<double> barsX = [];
+      var tempX = 0.0;
+      final barsX = <double>[];
       barGroup.barRods.asMap().forEach((barIndex, barRod) {
-        final double widthHalf = barRod.width / 2;
+        final widthHalf = barRod.width / 2;
         barsX.add(groupX - (barGroup.width / 2) + tempX + widthHalf);
         tempX += barRod.width + barGroup.barsSpace;
       });
@@ -182,18 +183,23 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
     return groupBarsPosition;
   }
 
-  void _drawBars(Canvas canvas, Size viewSize, List<_GroupBarsPosition> groupBarsPosition) {
-    final Size drawSize = getChartUsableDrawSize(viewSize);
+  void _drawBars(
+    CanvasWrapper canvasWrapper,
+    List<_GroupBarsPosition> groupBarsPosition,
+    PaintHolder<BarChartData> holder,
+  ) {
+    final data = holder.data;
+    final viewSize = canvasWrapper.size;
+    final drawSize = getChartUsableDrawSize(viewSize, holder);
 
-    for (int i = 0; i < data.barGroups.length; i++) {
+    for (var i = 0; i < data.barGroups.length; i++) {
       final barGroup = data.barGroups[i];
-      for (int j = 0; j < barGroup.barRods.length; j++) {
+      for (var j = 0; j < barGroup.barRods.length; j++) {
         final barRod = barGroup.barRods[j];
-        final double widthHalf = barRod.width / 2;
-        final BorderRadius borderRadius =
-            barRod.borderRadius ?? BorderRadius.circular(barRod.width / 2);
+        final widthHalf = barRod.width / 2;
+        final borderRadius = barRod.borderRadius ?? BorderRadius.circular(barRod.width / 2);
 
-        final double x = groupBarsPosition[i].barsX[j];
+        final x = groupBarsPosition[i].barsX[j];
 
         final left = x - widthHalf;
         final right = x + widthHalf;
@@ -206,8 +212,9 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
         if (barRod.backDrawRodData.show && barRod.backDrawRodData.y != 0) {
           if (barRod.backDrawRodData.y > 0) {
             // positive
-            final bottom = getPixelY(0, drawSize);
-            final top = min(getPixelY(barRod.backDrawRodData.y, drawSize), bottom - cornerHeight);
+            final bottom = getPixelY(max(data.minY, 0), drawSize, holder);
+            final top =
+                min(getPixelY(barRod.backDrawRodData.y, drawSize, holder), bottom - cornerHeight);
 
             barRRect = RRect.fromLTRBAndCorners(left, top, right, bottom,
                 topLeft: borderRadius.topLeft,
@@ -216,8 +223,9 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
                 bottomRight: borderRadius.bottomRight);
           } else {
             // negative
-            final top = getPixelY(0, drawSize);
-            final bottom = max(getPixelY(barRod.backDrawRodData.y, drawSize), top + cornerHeight);
+            final top = getPixelY(min(data.maxY, 0), drawSize, holder);
+            final bottom =
+                max(getPixelY(barRod.backDrawRodData.y, drawSize, holder), top + cornerHeight);
 
             barRRect = RRect.fromLTRBAndCorners(left, top, right, bottom,
                 topLeft: borderRadius.topLeft,
@@ -226,16 +234,36 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
                 bottomRight: borderRadius.bottomRight);
           }
 
-          _barPaint.color = barRod.backDrawRodData.color;
-          canvas.drawRRect(barRRect, _barPaint);
+          if (barRod.backDrawRodData.colors.length == 1) {
+            _barPaint.color = barRod.backDrawRodData.colors[0];
+            _barPaint.shader = null;
+          } else {
+            final from = barRod.backDrawRodData.gradientFrom;
+            final to = barRod.backDrawRodData.gradientTo;
+
+            _barPaint.shader = ui.Gradient.linear(
+              Offset(
+                getLeftOffsetDrawSize(holder) + (drawSize.width * from.dx),
+                getTopOffsetDrawSize(holder) + (drawSize.height * from.dy),
+              ),
+              Offset(
+                getLeftOffsetDrawSize(holder) + (drawSize.width * to.dx),
+                getTopOffsetDrawSize(holder) + (drawSize.height * to.dy),
+              ),
+              barRod.backDrawRodData.colors,
+              barRod.backDrawRodData.getSafeColorStops(),
+            );
+          }
+
+          canvasWrapper.drawRRect(barRRect, _barPaint);
         }
 
         // draw Main Rod
         if (barRod.y != 0) {
           if (barRod.y > 0) {
             // positive
-            final bottom = getPixelY(0, drawSize);
-            final top = min(getPixelY(barRod.y, drawSize), bottom - cornerHeight);
+            final bottom = getPixelY(max(data.minY, 0), drawSize, holder);
+            final top = min(getPixelY(barRod.y, drawSize, holder), bottom - cornerHeight);
 
             barRRect = RRect.fromLTRBAndCorners(left, top, right, bottom,
                 topLeft: borderRadius.topLeft,
@@ -244,8 +272,8 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
                 bottomRight: borderRadius.bottomRight);
           } else {
             // negative
-            final top = getPixelY(0, drawSize);
-            final bottom = max(getPixelY(barRod.y, drawSize), top + cornerHeight);
+            final top = getPixelY(min(data.maxY, 0), drawSize, holder);
+            final bottom = max(getPixelY(barRod.y, drawSize, holder), top + cornerHeight);
 
             barRRect = RRect.fromLTRBAndCorners(left, top, right, bottom,
                 topLeft: borderRadius.topLeft,
@@ -253,22 +281,40 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
                 bottomLeft: borderRadius.bottomLeft,
                 bottomRight: borderRadius.bottomRight);
           }
+          if (barRod.colors.length == 1) {
+            _barPaint.color = barRod.colors[0];
+            _barPaint.shader = null;
+          } else {
+            final from = barRod.gradientFrom;
+            final to = barRod.gradientTo;
 
-          _barPaint.color = barRod.color;
-          canvas.drawRRect(barRRect, _barPaint);
+            _barPaint.shader = ui.Gradient.linear(
+              Offset(
+                getLeftOffsetDrawSize(holder) + (drawSize.width * from.dx),
+                getTopOffsetDrawSize(holder) + (drawSize.height * from.dy),
+              ),
+              Offset(
+                getLeftOffsetDrawSize(holder) + (drawSize.width * to.dx),
+                getTopOffsetDrawSize(holder) + (drawSize.height * to.dy),
+              ),
+              barRod.colors,
+              barRod.getSafeColorStops(),
+            );
+          }
+          canvasWrapper.drawRRect(barRRect, _barPaint);
 
           // draw rod stack
-          if (barRod.rodStackItem != null && barRod.rodStackItem.isNotEmpty) {
-            for (int i = 0; i < barRod.rodStackItem.length; i++) {
-              final stackItem = barRod.rodStackItem[i];
-              final stackFromY = getPixelY(stackItem.fromY, drawSize);
-              final stackToY = getPixelY(stackItem.toY, drawSize);
+          if (barRod.rodStackItems.isNotEmpty) {
+            for (var i = 0; i < barRod.rodStackItems.length; i++) {
+              final stackItem = barRod.rodStackItems[i];
+              final stackFromY = getPixelY(stackItem.fromY, drawSize, holder);
+              final stackToY = getPixelY(stackItem.toY, drawSize, holder);
 
               _barPaint.color = stackItem.color;
-              canvas.save();
-              canvas.clipRect(Rect.fromLTRB(left, stackToY, right, stackFromY));
-              canvas.drawRRect(barRRect, _barPaint);
-              canvas.restore();
+              canvasWrapper.save();
+              canvasWrapper.clipRect(Rect.fromLTRB(left, stackToY, right, stackFromY));
+              canvasWrapper.drawRRect(barRRect, _barPaint);
+              canvasWrapper.restore();
             }
           }
         }
@@ -276,118 +322,173 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
     }
   }
 
-  void _drawTitles(Canvas canvas, Size viewSize, List<_GroupBarsPosition> groupBarsPosition) {
+  void _drawTitles(
+    CanvasWrapper canvasWrapper,
+    List<_GroupBarsPosition> groupBarsPosition,
+    PaintHolder<BarChartData> holder,
+  ) {
+    final data = holder.data;
+    final targetData = holder.targetData;
     if (!targetData.titlesData.show) {
       return;
     }
-    final Size drawSize = getChartUsableDrawSize(viewSize);
+    final viewSize = canvasWrapper.size;
+    final drawSize = getChartUsableDrawSize(viewSize, holder);
 
     // Left Titles
     final leftTitles = targetData.titlesData.leftTitles;
+    final leftInterval =
+        leftTitles.interval ?? getEfficientInterval(viewSize.height, data.verticalDiff);
     if (leftTitles.showTitles) {
-      double verticalSeek = data.minY;
+      var verticalSeek = data.minY;
       while (verticalSeek <= data.maxY) {
-        double x = 0 + getLeftOffsetDrawSize();
-        double y = getPixelY(verticalSeek, drawSize);
+        if (leftTitles.checkToShowTitle(
+            data.minY, data.maxY, leftTitles, leftInterval, verticalSeek)) {
+          var x = 0 + getLeftOffsetDrawSize(holder);
+          var y = getPixelY(verticalSeek, drawSize, holder);
 
-        final String text = leftTitles.getTitles(verticalSeek);
+          final text = leftTitles.getTitles(verticalSeek);
 
-        final TextSpan span = TextSpan(style: leftTitles.textStyle, text: text);
-        final TextPainter tp = TextPainter(
+          final span = TextSpan(style: leftTitles.getTextStyles(verticalSeek), text: text);
+          final tp = TextPainter(
+              text: span,
+              textAlign: TextAlign.center,
+              textDirection: TextDirection.ltr,
+              textScaleFactor: holder.textScale);
+          tp.layout(maxWidth: getExtraNeededHorizontalSpace(holder));
+          x -= tp.width + leftTitles.margin;
+          y -= tp.height / 2;
+          canvasWrapper.save();
+          canvasWrapper.translate(x + tp.width / 2, y + tp.height / 2);
+          canvasWrapper.rotate(radians(leftTitles.rotateAngle));
+          canvasWrapper.translate(-(x + tp.width / 2), -(y + tp.height / 2));
+          y -= translateRotatedPosition(tp.width, leftTitles.rotateAngle);
+          canvasWrapper.drawText(tp, Offset(x, y));
+          canvasWrapper.restore();
+        }
+        if (data.maxY - verticalSeek < leftInterval && data.maxY != verticalSeek) {
+          verticalSeek = data.maxY;
+        } else {
+          verticalSeek += leftInterval;
+        }
+      }
+    }
+
+    // Top Titles
+    final topTitles = targetData.titlesData.topTitles;
+    if (topTitles.showTitles) {
+      for (var index = 0; index < groupBarsPosition.length; index++) {
+        final groupBarPos = groupBarsPosition[index];
+
+        final xValue = data.barGroups[index].x.toDouble();
+        final text = topTitles.getTitles(xValue);
+        final span = TextSpan(style: topTitles.getTextStyles(xValue), text: text);
+        final tp = TextPainter(
             text: span,
             textAlign: TextAlign.center,
             textDirection: TextDirection.ltr,
-            textScaleFactor: textScale);
-        tp.layout(maxWidth: getExtraNeededHorizontalSpace());
-        x -= tp.width + leftTitles.margin;
-        y -= tp.height / 2;
-        canvas.save();
-        canvas.translate(x + tp.width / 2, y + tp.height / 2);
-        canvas.rotate(radians(leftTitles.rotateAngle));
-        canvas.translate(-(x + tp.width / 2), -(y + tp.height / 2));
-        y -= translateRotatedPosition(tp.width, leftTitles.rotateAngle);
-        tp.paint(canvas, Offset(x, y));
-        canvas.restore();
+            textScaleFactor: holder.textScale);
+        tp.layout();
+        var x = groupBarPos.groupX;
+        const y = 0.0;
 
-        verticalSeek += leftTitles.interval;
+        x -= tp.width / 2;
+        canvasWrapper.save();
+        canvasWrapper.translate(x + tp.width / 2, y + tp.height / 2);
+        canvasWrapper.rotate(radians(topTitles.rotateAngle));
+        canvasWrapper.translate(-(x + tp.width / 2), -(y + tp.height / 2));
+        x += translateRotatedPosition(tp.width, topTitles.rotateAngle);
+        canvasWrapper.drawText(tp, Offset(x, y));
+        canvasWrapper.restore();
       }
     }
 
     // Right Titles
     final rightTitles = targetData.titlesData.rightTitles;
+    final rightInterval =
+        rightTitles.interval ?? getEfficientInterval(viewSize.height, data.verticalDiff);
     if (rightTitles.showTitles) {
-      double verticalSeek = data.minY;
+      var verticalSeek = data.minY;
       while (verticalSeek <= data.maxY) {
-        double x = drawSize.width + getLeftOffsetDrawSize();
-        double y = getPixelY(verticalSeek, drawSize);
+        if (rightTitles.checkToShowTitle(
+            data.minY, data.maxY, rightTitles, rightInterval, verticalSeek)) {
+          var x = drawSize.width + getLeftOffsetDrawSize(holder);
+          var y = getPixelY(verticalSeek, drawSize, holder);
 
-        final String text = rightTitles.getTitles(verticalSeek);
+          final text = rightTitles.getTitles(verticalSeek);
 
-        final TextSpan span = TextSpan(style: rightTitles.textStyle, text: text);
-        final TextPainter tp = TextPainter(
-            text: span,
-            textAlign: TextAlign.center,
-            textDirection: TextDirection.ltr,
-            textScaleFactor: textScale);
-        tp.layout(maxWidth: getExtraNeededHorizontalSpace());
-        x += rightTitles.margin;
-        y -= tp.height / 2;
-        canvas.save();
-        canvas.translate(x + tp.width / 2, y + tp.height / 2);
-        canvas.rotate(radians(rightTitles.rotateAngle));
-        canvas.translate(-(x + tp.width / 2), -(y + tp.height / 2));
-        y += translateRotatedPosition(tp.width, leftTitles.rotateAngle);
-        tp.paint(canvas, Offset(x, y));
-        canvas.restore();
-
-        verticalSeek += rightTitles.interval;
+          final span = TextSpan(style: rightTitles.getTextStyles(verticalSeek), text: text);
+          final tp = TextPainter(
+              text: span,
+              textAlign: TextAlign.center,
+              textDirection: TextDirection.ltr,
+              textScaleFactor: holder.textScale);
+          tp.layout(maxWidth: getExtraNeededHorizontalSpace(holder));
+          x += rightTitles.margin;
+          y -= tp.height / 2;
+          canvasWrapper.save();
+          canvasWrapper.translate(x + tp.width / 2, y + tp.height / 2);
+          canvasWrapper.rotate(radians(rightTitles.rotateAngle));
+          canvasWrapper.translate(-(x + tp.width / 2), -(y + tp.height / 2));
+          y += translateRotatedPosition(tp.width, leftTitles.rotateAngle);
+          canvasWrapper.drawText(tp, Offset(x, y));
+          canvasWrapper.restore();
+        }
+        if (data.maxY - verticalSeek < rightInterval && data.maxY != verticalSeek) {
+          verticalSeek = data.maxY;
+        } else {
+          verticalSeek += rightInterval;
+        }
       }
     }
 
     // Bottom titles
     final bottomTitles = targetData.titlesData.bottomTitles;
     if (bottomTitles.showTitles) {
-      for (int index = 0; index < groupBarsPosition.length; index++) {
-        final _GroupBarsPosition groupBarPos = groupBarsPosition[index];
+      for (var index = 0; index < groupBarsPosition.length; index++) {
+        final groupBarPos = groupBarsPosition[index];
 
-        final String text = bottomTitles.getTitles(index.toDouble());
-        final TextSpan span = TextSpan(style: bottomTitles.textStyle, text: text);
-        final TextPainter tp = TextPainter(
+        final xValue = data.barGroups[index].x.toDouble();
+        final text = bottomTitles.getTitles(xValue);
+        // ignore: omit_local_variable_types
+        final span = TextSpan(style: bottomTitles.getTextStyles(xValue), text: text);
+        final tp = TextPainter(
             text: span,
             textAlign: TextAlign.center,
             textDirection: TextDirection.ltr,
-            textScaleFactor: textScale);
+            textScaleFactor: holder.textScale);
         tp.layout();
-        double x = groupBarPos.groupX;
-        final double y = drawSize.height + getTopOffsetDrawSize() + bottomTitles.margin;
+        var x = groupBarPos.groupX;
+        final y = drawSize.height + getTopOffsetDrawSize(holder) + bottomTitles.margin;
 
         x -= tp.width / 2;
-        canvas.save();
-        canvas.translate(x + tp.width / 2, y + tp.height / 2);
-        canvas.rotate(radians(bottomTitles.rotateAngle));
-        canvas.translate(-(x + tp.width / 2), -(y + tp.height / 2));
+        canvasWrapper.save();
+        canvasWrapper.translate(x + tp.width / 2, y + tp.height / 2);
+        canvasWrapper.rotate(radians(bottomTitles.rotateAngle));
+        canvasWrapper.translate(-(x + tp.width / 2), -(y + tp.height / 2));
         x += translateRotatedPosition(tp.width, bottomTitles.rotateAngle);
-        tp.paint(canvas, Offset(x, y));
-        canvas.restore();
+        canvasWrapper.drawText(tp, Offset(x, y));
+        canvasWrapper.restore();
       }
     }
   }
 
   void _drawTouchTooltip(
-    Canvas canvas,
-    Size viewSize,
+    CanvasWrapper canvasWrapper,
     List<_GroupBarsPosition> groupPositions,
     BarTouchTooltipData tooltipData,
     BarChartGroupData showOnBarGroup,
     int barGroupIndex,
     BarChartRodData showOnRodData,
     int barRodIndex,
+    PaintHolder<BarChartData> holder,
   ) {
-    final Size chartUsableSize = getChartUsableDrawSize(viewSize);
+    final viewSize = canvasWrapper.size;
+    final chartUsableSize = getChartUsableDrawSize(viewSize, holder);
 
-    const double textsBelowMargin = 4;
+    const textsBelowMargin = 4;
 
-    final BarTooltipItem tooltipItem = tooltipData.getTooltipItem(
+    final tooltipItem = tooltipData.getTooltipItem(
       showOnBarGroup,
       barGroupIndex,
       showOnRodData,
@@ -398,16 +499,16 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
       return;
     }
 
-    final TextSpan span = TextSpan(style: tooltipItem.textStyle, text: tooltipItem.text);
-    final TextPainter tp = TextPainter(
+    final span = TextSpan(style: tooltipItem.textStyle, text: tooltipItem.text);
+    final tp = TextPainter(
         text: span,
-        textAlign: TextAlign.center,
+        textAlign: tooltipItem.textAlign,
         textDirection: TextDirection.ltr,
-        textScaleFactor: textScale);
+        textScaleFactor: holder.textScale);
     tp.layout(maxWidth: tooltipData.maxContentWidth);
 
     /// creating TextPainters to calculate the width and height of the tooltip
-    final TextPainter drawingTextPainter = tp;
+    final drawingTextPainter = tp;
 
     /// biggerWidth
     /// some texts maybe larger, then we should
@@ -416,27 +517,31 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
     /// sumTextsHeight
     /// sum up all Texts height, then we should
     /// draw the tooltip's height as tall as sumTextsHeight
-    final double textWidth = drawingTextPainter.width;
-    final double textHeight = drawingTextPainter.height + textsBelowMargin;
+    final textWidth = drawingTextPainter.width;
+    final textHeight = drawingTextPainter.height + textsBelowMargin;
 
     /// if we have multiple bar lines,
     /// there are more than one FlCandidate on touch area,
     /// we should get the most top FlSpot Offset to draw the tooltip on top of it
-    final Offset barOffset = Offset(
+    final barOffset = Offset(
       groupPositions[barGroupIndex].barsX[barRodIndex],
-      getPixelY(showOnRodData.y, chartUsableSize),
+      getPixelY(showOnRodData.y, chartUsableSize, holder),
     );
 
-    final isPositive = showOnRodData.y > 0;
+    final tooltipWidth = textWidth + tooltipData.tooltipPadding.horizontal;
+    final tooltipHeight = textHeight + tooltipData.tooltipPadding.vertical;
 
-    final double tooltipWidth = textWidth + tooltipData.tooltipPadding.horizontal;
-    final double tooltipHeight = textHeight + tooltipData.tooltipPadding.vertical;
-
-    final double tooltipTop = isPositive
-        ? barOffset.dy - tooltipHeight - tooltipData.tooltipBottomMargin
-        : barOffset.dy + tooltipData.tooltipBottomMargin;
+    final zeroY = getPixelY(0, chartUsableSize, holder);
+    final barTopY = min(zeroY, barOffset.dy);
+    final barBottomY = max(zeroY, barOffset.dy);
+    final drawTooltipOnTop = tooltipData.direction == TooltipDirection.top ||
+        (tooltipData.direction == TooltipDirection.auto && showOnRodData.y > 0);
+    final tooltipTop = drawTooltipOnTop
+        ? barTopY - tooltipHeight - tooltipData.tooltipMargin
+        : barBottomY + tooltipData.tooltipMargin;
 
     /// draw the background rect with rounded radius
+    // ignore: omit_local_variable_types
     Rect rect =
         Rect.fromLTWH(barOffset.dx - (tooltipWidth / 2), tooltipTop, tooltipWidth, tooltipHeight);
 
@@ -484,19 +589,19 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
       }
     }
 
-    final Radius radius = Radius.circular(tooltipData.tooltipRoundedRadius);
-    final RRect roundedRect = RRect.fromRectAndCorners(rect,
+    final radius = Radius.circular(tooltipData.tooltipRoundedRadius);
+    final roundedRect = RRect.fromRectAndCorners(rect,
         topLeft: radius, topRight: radius, bottomLeft: radius, bottomRight: radius);
     _bgTouchTooltipPaint.color = tooltipData.tooltipBgColor;
-    canvas.drawRRect(roundedRect, _bgTouchTooltipPaint);
+    canvasWrapper.drawRRect(roundedRect, _bgTouchTooltipPaint);
 
     /// draw the texts one by one in below of each other
-    final double top = tooltipData.tooltipPadding.top;
+    final top = tooltipData.tooltipPadding.top;
     final drawOffset = Offset(
       rect.center.dx - (tp.width / 2),
       rect.topCenter.dy + top,
     );
-    tp.paint(canvas, drawOffset);
+    canvasWrapper.drawText(tp, drawOffset);
   }
 
   /// We add our needed horizontal space to parent needed.
@@ -505,8 +610,9 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
   /// the left space is [getLeftOffsetDrawSize],
   /// and the whole space is [getExtraNeededHorizontalSpace]
   @override
-  double getExtraNeededHorizontalSpace() {
-    double sum = super.getExtraNeededHorizontalSpace();
+  double getExtraNeededHorizontalSpace(PaintHolder<BarChartData> holder) {
+    final data = holder.data;
+    var sum = super.getExtraNeededHorizontalSpace(holder);
     if (data.titlesData.show) {
       final leftSide = data.titlesData.leftTitles;
       if (leftSide.showTitles) {
@@ -527,12 +633,18 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
   /// the top space is [getTopOffsetDrawSize()],
   /// and the whole space is [getExtraNeededVerticalSpace]
   @override
-  double getExtraNeededVerticalSpace() {
-    double sum = super.getExtraNeededVerticalSpace();
+  double getExtraNeededVerticalSpace(PaintHolder<BarChartData> holder) {
+    final data = holder.data;
+    var sum = super.getExtraNeededVerticalSpace(holder);
     if (data.titlesData.show) {
       final bottomSide = data.titlesData.bottomTitles;
       if (bottomSide.showTitles) {
         sum += bottomSide.reservedSize + bottomSide.margin;
+      }
+
+      final topSide = data.titlesData.topTitles;
+      if (topSide.showTitles) {
+        sum += topSide.reservedSize + topSide.margin;
       }
     }
     return sum;
@@ -542,8 +654,9 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
   /// maybe we want to show both left and right titles,
   /// then just the left titles will effect on this function.
   @override
-  double getLeftOffsetDrawSize() {
-    var sum = super.getLeftOffsetDrawSize();
+  double getLeftOffsetDrawSize(PaintHolder<BarChartData> holder) {
+    final data = holder.data;
+    var sum = super.getLeftOffsetDrawSize(holder);
 
     final leftTitles = data.titlesData.leftTitles;
     if (data.titlesData.show && leftTitles.showTitles) {
@@ -553,57 +666,73 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
     return sum;
   }
 
-  /// Makes a [BarTouchResponse] based on the provided [FlTouchInput]
-  ///
-  /// Processes [FlTouchInput.getOffset] and checks
-  /// the elements of the chart that are near the offset,
-  /// then makes a [BarTouchResponse] from the elements that has been touched.
+  /// calculate top offset for draw the chart,
+  /// maybe we want to show both top and bottom titles,
+  /// then just the top titles will effect on this function.
   @override
-  BarTouchResponse handleTouch(FlTouchInput touchInput, Size size) {
-    final BarTouchedSpot touchedSpot =
-        _getNearestTouchedSpot(size, touchInput.getOffset(), _groupBarsPosition);
-    return BarTouchResponse(touchedSpot, touchInput);
-  }
+  double getTopOffsetDrawSize(PaintHolder<BarChartData> holder) {
+    final data = holder.data;
+    var sum = super.getTopOffsetDrawSize(holder);
 
-  /// find the nearest spot base on the touched offset
-  BarTouchedSpot _getNearestTouchedSpot(
-      Size viewSize, Offset touchedPoint, List<_GroupBarsPosition> groupBarsPosition) {
-    if (groupBarsPosition == null) {
-      final List<double> groupsX = _calculateGroupsX(viewSize, data.barGroups, data.alignment);
-      groupBarsPosition = _calculateGroupAndBarsPosition(viewSize, groupsX, data.barGroups);
+    final topTitles = data.titlesData.topTitles;
+    if (data.titlesData.show && topTitles.showTitles) {
+      sum += topTitles.reservedSize + topTitles.margin;
     }
 
-    final Size chartViewSize = getChartUsableDrawSize(viewSize);
+    return sum;
+  }
+
+  /// Makes a [BarTouchedSpot] based on the provided [touchInput]
+  ///
+  /// Processes [PointerEvent.localPosition] and checks
+  /// the elements of the chart that are near the offset,
+  /// then makes a [BarTouchedSpot] from the elements that has been touched.
+  ///
+  /// Returns null if finds nothing!
+  BarTouchedSpot? handleTouch(
+    PointerEvent touchInput,
+    Size viewSize,
+    PaintHolder<BarChartData> holder,
+  ) {
+    final data = holder.data;
+    final targetData = holder.targetData;
+    final touchedPoint = touchInput.localPosition;
+    if (_groupBarsPosition == null) {
+      final groupsX = _calculateGroupsX(viewSize, data.barGroups, data.alignment, holder);
+      _groupBarsPosition = _calculateGroupAndBarsPosition(viewSize, groupsX, data.barGroups);
+    }
+
+    final chartViewSize = getChartUsableDrawSize(viewSize, holder);
 
     /// Find the nearest barRod
-    for (int i = 0; i < groupBarsPosition.length; i++) {
-      final _GroupBarsPosition groupBarPos = groupBarsPosition[i];
-      for (int j = 0; j < groupBarPos.barsX.length; j++) {
-        final double barX = groupBarPos.barsX[j];
-        final double barWidth = targetData.barGroups[i].barRods[j].width;
-        final double halfBarWidth = barWidth / 2;
+    for (var i = 0; i < _groupBarsPosition!.length; i++) {
+      final groupBarPos = _groupBarsPosition![i];
+      for (var j = 0; j < groupBarPos.barsX.length; j++) {
+        final barX = groupBarPos.barsX[j];
+        final barWidth = targetData.barGroups[i].barRods[j].width;
+        final halfBarWidth = barWidth / 2;
 
         double barTopY;
         double barBotY;
 
-        final bool isPositive = targetData.barGroups[i].barRods[j].y > 0;
+        final isPositive = targetData.barGroups[i].barRods[j].y > 0;
         if (isPositive) {
-          barTopY = getPixelY(targetData.barGroups[i].barRods[j].y, chartViewSize);
-          barBotY = getPixelY(0, chartViewSize);
+          barTopY = getPixelY(targetData.barGroups[i].barRods[j].y, chartViewSize, holder);
+          barBotY = getPixelY(0, chartViewSize, holder);
         } else {
-          barTopY = getPixelY(0, chartViewSize);
-          barBotY = getPixelY(targetData.barGroups[i].barRods[j].y, chartViewSize);
+          barTopY = getPixelY(0, chartViewSize, holder);
+          barBotY = getPixelY(targetData.barGroups[i].barRods[j].y, chartViewSize, holder);
         }
 
-        final double backDrawBarY =
-            getPixelY(targetData.barGroups[i].barRods[j].backDrawRodData.y, chartViewSize);
-        final EdgeInsets touchExtraThreshold = targetData.barTouchData.touchExtraThreshold;
+        final backDrawBarY =
+            getPixelY(targetData.barGroups[i].barRods[j].backDrawRodData.y, chartViewSize, holder);
+        final touchExtraThreshold = targetData.barTouchData.touchExtraThreshold;
 
-        final bool isXInTouchBounds =
+        final isXInTouchBounds =
             (touchedPoint.dx <= barX + halfBarWidth + touchExtraThreshold.right) &&
                 (touchedPoint.dx >= barX - halfBarWidth - touchExtraThreshold.left);
 
-        final bool isYInBarBounds = (touchedPoint.dy <= barBotY + touchExtraThreshold.bottom) &&
+        final isYInBarBounds = (touchedPoint.dy <= barBotY + touchExtraThreshold.bottom) &&
             (touchedPoint.dy >= barTopY - touchExtraThreshold.top);
 
         bool isYInBarBackDrawBounds;
@@ -615,7 +744,7 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
               (touchedPoint.dy <= backDrawBarY + touchExtraThreshold.bottom);
         }
 
-        final bool isYInTouchBounds =
+        final isYInTouchBounds =
             (targetData.barTouchData.allowTouchBarBackDraw && isYInBarBackDrawBounds) ||
                 isYInBarBounds;
 
@@ -623,22 +752,29 @@ class BarChartPainter extends AxisChartPainter<BarChartData> with TouchHandler<B
           final nearestGroup = targetData.barGroups[i];
           final nearestBarRod = nearestGroup.barRods[j];
           final nearestSpot = FlSpot(nearestGroup.x.toDouble(), nearestBarRod.y);
-          final nearestSpotPos = Offset(barX, getPixelY(nearestSpot.y, chartViewSize));
+          final nearestSpotPos = Offset(barX, getPixelY(nearestSpot.y, chartViewSize, holder));
 
-          return BarTouchedSpot(nearestGroup, i, nearestBarRod, j, nearestSpot, nearestSpotPos);
+          var touchedStackIndex = -1;
+          BarChartRodStackItem? touchedStack;
+          for (var stackIndex = 0; stackIndex < nearestBarRod.rodStackItems.length; stackIndex++) {
+            final stackItem = nearestBarRod.rodStackItems[stackIndex];
+            final fromPixel = getPixelY(stackItem.fromY, chartViewSize, holder);
+            final toPixel = getPixelY(stackItem.toY, chartViewSize, holder);
+            if (touchedPoint.dy <= fromPixel && touchedPoint.dy >= toPixel) {
+              touchedStackIndex = stackIndex;
+              touchedStack = stackItem;
+              break;
+            }
+          }
+
+          return BarTouchedSpot(nearestGroup, i, nearestBarRod, j, touchedStack, touchedStackIndex,
+              nearestSpot, nearestSpotPos);
         }
       }
     }
 
     return null;
   }
-
-  /// Determines should it redraw the chart or not.
-  ///
-  /// If there is a change in the [BarChartData],
-  /// [BarChartPainter] should repaint itself.
-  @override
-  bool shouldRepaint(BarChartPainter oldDelegate) => oldDelegate.data != data;
 }
 
 class _GroupBarsPosition {
