@@ -1,23 +1,26 @@
-import 'package:fl_chart/src/chart/base/base_chart/base_chart_painter.dart';
-import 'package:fl_chart/src/chart/base/base_chart/touch_input.dart';
-import 'package:fl_chart/src/chart/pie_chart/pie_chart_painter.dart';
-import 'package:fl_chart/src/utils/utils.dart';
+import 'package:fl_chart/src/chart/pie_chart/pie_chart_renderer.dart';
 import 'package:flutter/material.dart';
 
 import 'pie_chart_data.dart';
 
 /// Renders a pie chart as a widget, using provided [PieChartData].
 class PieChart extends ImplicitlyAnimatedWidget {
+  /// Default duration to reuse externally.
+  static const defaultDuration = Duration(milliseconds: 150);
+
   /// Determines how the [PieChart] should be look like.
   final PieChartData data;
 
   /// [data] determines how the [PieChart] should be look like,
   /// when you make any change in the [PieChartData], it updates
   /// new values with animation, and duration is [swapAnimationDuration].
+  /// also you can change the [swapAnimationCurve]
+  /// which default is [Curves.linear].
   const PieChart(
     this.data, {
-    Duration swapAnimationDuration = const Duration(milliseconds: 150),
-  }) : super(duration: swapAnimationDuration);
+    Duration swapAnimationDuration = defaultDuration,
+    Curve swapAnimationCurve = Curves.linear,
+  }) : super(duration: swapAnimationDuration, curve: swapAnimationCurve);
 
   /// Creates a [_PieChartState]
   @override
@@ -25,132 +28,38 @@ class PieChart extends ImplicitlyAnimatedWidget {
 }
 
 class _PieChartState extends AnimatedWidgetBaseState<PieChart> {
-  /// we handle under the hood animations (implicit animations) via this tween,
+  /// We handle under the hood animations (implicit animations) via this tween,
   /// it lerps between the old [PieChartData] to the new one.
-  PieChartDataTween _pieChartDataTween;
+  PieChartDataTween? _pieChartDataTween;
 
-  /// this is used to map the touch events to [PieTouchResponse]
-  TouchHandler _touchHandler;
-
-  /// this is used to retrieve the chart size to handle the touches
-  final GlobalKey _chartKey = GlobalKey();
+  @override
+  void initState() {
+    /// Make sure that [_widgetsPositionHandler] is updated.
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
+      if (mounted) {
+        setState(() {});
+      }
+    });
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
-    final PieChartData showingData = _getData();
-    final PieTouchData touchData = showingData.pieTouchData;
+    final showingData = _getData();
 
+    /// Wr wrapped our chart with [GestureDetector], and onLongPressStart callback.
+    /// because we wanted to lock the widget from being scrolled when user long presses on it.
+    /// If we found a solution for solve this issue, then we can remove this undoubtedly.
     return GestureDetector(
-      onLongPressStart: (d) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlLongPressStart(d.localPosition), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onLongPressEnd: (d) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlLongPressEnd(d.localPosition), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onLongPressMoveUpdate: (d) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlLongPressMoveUpdate(d.localPosition), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onPanCancel: () {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response = _touchHandler?.handleTouch(
-            FlPanEnd(Offset.zero, const Velocity(pixelsPerSecond: Offset.zero)), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onPanEnd: (DragEndDetails details) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlPanEnd(Offset.zero, details.velocity), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onPanDown: (DragDownDetails details) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlPanStart(details.localPosition), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      onPanUpdate: (DragUpdateDetails details) {
-        final Size chartSize = _getChartSize();
-        if (chartSize == null) {
-          return;
-        }
-
-        final PieTouchResponse response =
-            _touchHandler?.handleTouch(FlPanMoveUpdate(details.localPosition), chartSize);
-        if (_canHandleTouch(response, touchData)) {
-          touchData.touchCallback(response);
-        }
-      },
-      child: CustomPaint(
-        key: _chartKey,
-        size: getDefaultSize(MediaQuery.of(context).size),
-        painter: PieChartPainter(
-          _pieChartDataTween.evaluate(animation),
-          showingData,
-          (touchHandler) {
-            setState(() {
-              _touchHandler = touchHandler;
-            });
-          },
-          textScale: MediaQuery.of(context).textScaleFactor,
-        ),
+      onLongPressStart: (details) {},
+      child: PieChartLeaf(
+        data: _pieChartDataTween!.evaluate(animation),
+        targetData: showingData,
+        touchCallback: (response) {
+          showingData.pieTouchData.touchCallback?.call(response);
+        },
       ),
     );
-  }
-
-  bool _canHandleTouch(PieTouchResponse response, PieTouchData touchData) {
-    return response != null && touchData != null && touchData.touchCallback != null;
-  }
-
-  Size _getChartSize() {
-    final RenderBox containerRenderBox = _chartKey.currentContext?.findRenderObject();
-    if (containerRenderBox != null && containerRenderBox.hasSize) {
-      return containerRenderBox.size;
-    }
-    return null;
   }
 
   /// if builtIn touches are enabled, we should recreate our [pieChartData]
@@ -164,7 +73,46 @@ class _PieChartState extends AnimatedWidgetBaseState<PieChart> {
     _pieChartDataTween = visitor(
       _pieChartDataTween,
       widget.data,
-      (dynamic value) => PieChartDataTween(begin: value),
-    );
+      (dynamic value) => PieChartDataTween(begin: value, end: widget.data),
+    ) as PieChartDataTween;
+  }
+}
+
+/// Positions the badge widgets on their respective sections.
+class BadgeWidgetsDelegate extends MultiChildLayoutDelegate {
+  final int badgeWidgetsCount;
+  final Map<int, Offset> badgeWidgetsOffsets;
+
+  BadgeWidgetsDelegate({
+    required this.badgeWidgetsCount,
+    required this.badgeWidgetsOffsets,
+  });
+
+  @override
+  void performLayout(Size size) {
+    for (var index = 0; index < badgeWidgetsCount; index++) {
+      final _key = badgeWidgetsOffsets.keys.elementAt(index);
+
+      final _size = layoutChild(
+        _key,
+        BoxConstraints(
+          maxWidth: size.width,
+          maxHeight: size.height,
+        ),
+      );
+
+      positionChild(
+        _key,
+        Offset(
+          badgeWidgetsOffsets[_key]!.dx - (_size.width / 2),
+          badgeWidgetsOffsets[_key]!.dy - (_size.height / 2),
+        ),
+      );
+    }
+  }
+
+  @override
+  bool shouldRelayout(BadgeWidgetsDelegate oldDelegate) {
+    return oldDelegate.badgeWidgetsOffsets != badgeWidgetsOffsets;
   }
 }
