@@ -1,9 +1,9 @@
-import 'package:example/line_chart/samples/line_chart_sample7.dart';
 import 'package:flutter/material.dart';
 
 import 'samples/line_chart_sample3.dart';
 import 'samples/line_chart_sample4.dart';
 import 'samples/line_chart_sample5.dart';
+import 'samples/line_chart_sample7.dart';
 import 'samples/line_chart_sample8.dart';
 
 class LineChartPage2 extends StatelessWidget {
