@@ -1,7 +1,8 @@
-import 'package:example/bar_chart/samples/bar_chart_sample3.dart';
-import 'package:example/bar_chart/samples/bar_chart_sample4.dart';
-import 'package:example/bar_chart/samples/bar_chart_sample5.dart';
 import 'package:flutter/material.dart';
+
+import '../bar_chart/samples/bar_chart_sample3.dart';
+import '../bar_chart/samples/bar_chart_sample4.dart';
+import '../bar_chart/samples/bar_chart_sample5.dart';
 
 class LineChartPage4 extends StatelessWidget {
   @override
