@@ -1,16 +1,24 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/gestures.dart';
 
-class LineChartSample3 extends StatelessWidget {
-  final weekDays = [
-    'Sat',
-    'Sun',
-    'Mon',
-    'Tue',
-    'Wed',
-    'Thu',
-    'Fri',
-  ];
+class LineChartSample3 extends StatefulWidget {
+  final weekDays = ['Sat', 'Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri'];
+
+  final List<double> yValues = [1.3, 1, 1.8, 1.5, 2.2, 1.8, 3];
+
+  @override
+  State createState() => _LineChartSample3State();
+}
+
+class _LineChartSample3State extends State<LineChartSample3> {
+  double touchedValue;
+
+  @override
+  void initState() {
+    touchedValue = -1;
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -53,10 +61,22 @@ class LineChartSample3 extends StatelessWidget {
                       return TouchedSpotIndicatorData(
                         FlLine(color: Colors.blue, strokeWidth: 4),
                         FlDotData(
-                          dotSize: 8,
-                          strokeWidth: 5,
-                          getDotColor: (spot, percent, barData) => Colors.white,
-                          getStrokeColor: (spot, percent, barData) => Colors.deepOrange,
+                          getDotPainter: (spot, percent, barData, index) {
+                            if (index % 2 == 0) {
+                              return FlDotCirclePainter(
+                                  radius: 8,
+                                  color: Colors.white,
+                                  strokeWidth: 5,
+                                  strokeColor: Colors.deepOrange);
+                            } else {
+                              return FlDotSquarePainter(
+                                size: 16,
+                                color: Colors.white,
+                                strokeWidth: 5,
+                                strokeColor: Colors.deepOrange,
+                              );
+                            }
+                          },
                         ),
                       );
                     }).toList();
@@ -71,11 +91,34 @@ class LineChartSample3 extends StatelessWidget {
                           }
 
                           return LineTooltipItem(
-                            '${weekDays[flSpot.x.toInt()]} \n${flSpot.y} k calories',
+                            '${widget.weekDays[flSpot.x.toInt()]} \n${flSpot.y} k calories',
                             const TextStyle(color: Colors.white),
                           );
                         }).toList();
-                      })),
+                      }),
+                  touchCallback: (LineTouchResponse lineTouch) {
+                    final desiredTouch = lineTouch.touchInput is! PointerExitEvent &&
+                        lineTouch.touchInput is! PointerUpEvent;
+
+                    if (desiredTouch && lineTouch.lineBarSpots != null) {
+                      final value = lineTouch.lineBarSpots[0].x;
+
+                      if (value == 0 || value == 6) {
+                        setState(() {
+                          touchedValue = -1;
+                        });
+                        return null;
+                      }
+
+                      setState(() {
+                        touchedValue = value;
+                      });
+                    } else {
+                      setState(() {
+                        touchedValue = -1;
+                      });
+                    }
+                  }),
               extraLinesData: ExtraLinesData(horizontalLines: [
                 HorizontalLine(
                   y: 1.8,
@@ -87,15 +130,9 @@ class LineChartSample3 extends StatelessWidget {
               lineBarsData: [
                 LineChartBarData(
                   isStepLineChart: true,
-                  spots: [
-                    FlSpot(0, 1.3),
-                    FlSpot(1, 1),
-                    FlSpot(2, 1.8),
-                    FlSpot(3, 1.5),
-                    FlSpot(4, 2.2),
-                    FlSpot(5, 1.8),
-                    FlSpot(6, 3),
-                  ],
+                  spots: widget.yValues.asMap().entries.map((e) {
+                    return FlSpot(e.key.toDouble(), e.value);
+                  }).toList(),
                   isCurved: false,
                   barWidth: 4,
                   colors: [
@@ -127,10 +164,22 @@ class LineChartSample3 extends StatelessWidget {
                   ),
                   dotData: FlDotData(
                       show: true,
-                      getDotColor: (spot, percent, barData) => Colors.white,
-                      dotSize: 6,
-                      strokeWidth: 3,
-                      getStrokeColor: (spot, percent, barData) => Colors.deepOrange,
+                      getDotPainter: (spot, percent, barData, index) {
+                        if (index % 2 == 0) {
+                          return FlDotCirclePainter(
+                              radius: 6,
+                              color: Colors.white,
+                              strokeWidth: 3,
+                              strokeColor: Colors.deepOrange);
+                        } else {
+                          return FlDotSquarePainter(
+                            size: 12,
+                            color: Colors.white,
+                            strokeWidth: 3,
+                            strokeColor: Colors.deepOrange,
+                          );
+                        }
+                      },
                       checkToShowDot: (spot, barData) {
                         return spot.x != 0 && spot.x != 6;
                       }),
@@ -172,6 +221,7 @@ class LineChartSample3 extends StatelessWidget {
                 show: true,
                 leftTitles: SideTitles(
                   showTitles: true,
+                  reservedSize: 30,
                   getTitles: (value) {
                     switch (value.toInt()) {
                       case 0:
@@ -186,17 +236,20 @@ class LineChartSample3 extends StatelessWidget {
 
                     return '';
                   },
-                  textStyle: const TextStyle(color: Colors.black, fontSize: 10),
+                  getTextStyles: (value) => const TextStyle(color: Colors.black, fontSize: 10),
                 ),
                 bottomTitles: SideTitles(
                   showTitles: true,
                   getTitles: (value) {
-                    return weekDays[value.toInt()];
+                    return widget.weekDays[value.toInt()];
                   },
-                  textStyle: const TextStyle(
-                    color: Colors.deepOrange,
-                    fontWeight: FontWeight.bold,
-                  ),
+                  getTextStyles: (value) {
+                    final isTouched = value == touchedValue;
+                    return TextStyle(
+                      color: isTouched ? Colors.deepOrange : Colors.deepOrange.withOpacity(0.5),
+                      fontWeight: FontWeight.bold,
+                    );
+                  },
                 ),
               ),
             ),
