@@ -1,3 +1,6 @@
+import 'package:example/radar_chart/radar_chart_page.dart';
+import 'package:example/scatter_chart/scatter_chart_page.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'bar_chart/bar_chart_page.dart';
@@ -37,24 +40,69 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  int _currentPage = 0;
+
+  final _controller = PageController(initialPage: 0);
+  final _duration = Duration(milliseconds: 300);
+  final _curve = Curves.easeInOutCubic;
+  final _pages = [
+    LineChartPage(),
+    BarChartPage(),
+    BarChartPage2(),
+    PieChartPage(),
+    LineChartPage2(),
+    LineChartPage3(),
+    LineChartPage4(),
+    LineChartPage5(),
+    ScatterChartPage(),
+    RadarChartPage(),
+  ];
+
+  @override
+  void initState() {
+    _controller.addListener(() {
+      setState(() {
+        _currentPage = _controller.page.round();
+      });
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: SafeArea(
         child: PageView(
-          children: <Widget>[
-            LineChartPage(),
-            BarChartPage(),
-            BarChartPage2(),
-            PieChartPage(),
-            LineChartPage2(),
-            LineChartPage3(),
-            LineChartPage4(),
-            LineChartPage5(),
-            ScatterChartPage(),
-          ],
+          physics: kIsWeb ? NeverScrollableScrollPhysics() : AlwaysScrollableScrollPhysics(),
+          controller: _controller,
+          children: _pages,
         ),
       ),
+      bottomNavigationBar: kIsWeb
+          ? Container(
+              padding: EdgeInsets.all(16),
+              color: Colors.transparent,
+              child: Row(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Visibility(
+                    visible: _currentPage != 0,
+                    child: FloatingActionButton(
+                      onPressed: () => _controller.previousPage(duration: _duration, curve: _curve),
+                      child: Icon(Icons.chevron_left_rounded),
+                    ),
+                  ),
+                  Spacer(),
+                  Visibility(
+                    visible: _currentPage != _pages.length - 1,
+                    child: FloatingActionButton(
+                      onPressed: () => _controller.nextPage(duration: _duration, curve: _curve),
+                      child: Icon(Icons.chevron_right_rounded),
+                    ),
+                  ),
+                ],
+              ),
+            )
+          : null,
     );
   }
 }
