@@ -1,6 +1,6 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-
+import 'package:flutter/gestures.dart';
 import 'indicator.dart';
 
 class PieChartSample1 extends StatefulWidget {
@@ -66,11 +66,12 @@ class PieChartSample1State extends State {
                   PieChartData(
                       pieTouchData: PieTouchData(touchCallback: (pieTouchResponse) {
                         setState(() {
-                          if (pieTouchResponse.touchInput is FlLongPressEnd ||
-                              pieTouchResponse.touchInput is FlPanEnd) {
-                            touchedIndex = -1;
+                          final desiredTouch = pieTouchResponse.touchInput is! PointerExitEvent &&
+                              pieTouchResponse.touchInput is! PointerUpEvent;
+                          if (desiredTouch && pieTouchResponse.touchedSection != null) {
+                            touchedIndex = pieTouchResponse.touchedSection.touchedSectionIndex;
                           } else {
-                            touchedIndex = pieTouchResponse.touchedSectionIndex;
+                            touchedIndex = -1;
                           }
                         });
                       }),
@@ -78,7 +79,7 @@ class PieChartSample1State extends State {
                       borderData: FlBorderData(
                         show: false,
                       ),
-                      sectionsSpace: 12,
+                      sectionsSpace: 1,
                       centerSpaceRadius: 0,
                       sections: showingSections()),
                 ),
