@@ -13,8 +13,6 @@ class _ScatterChartSample2State extends State {
 
   List<int> selectedSpots = [];
 
-  int lastPanStartOnIndex = -1;
-
   @override
   Widget build(BuildContext context) {
     return AspectRatio(
@@ -99,21 +97,16 @@ class _ScatterChartSample2State extends State {
                 tooltipBgColor: Colors.black,
               ),
               touchCallback: (ScatterTouchResponse touchResponse) {
-                if (touchResponse.touchInput is FlPanStart) {
-                  lastPanStartOnIndex = touchResponse.touchedSpotIndex;
-                } else if (touchResponse.touchInput is FlPanEnd) {
-                  final FlPanEnd flPanEnd = touchResponse.touchInput;
-
-                  if (flPanEnd.velocity.pixelsPerSecond <= const Offset(4, 4)) {
-                    // Tap happened
-                    setState(() {
-                      if (selectedSpots.contains(lastPanStartOnIndex)) {
-                        selectedSpots.remove(lastPanStartOnIndex);
-                      } else {
-                        selectedSpots.add(lastPanStartOnIndex);
-                      }
-                    });
-                  }
+                if (touchResponse.clickHappened && touchResponse.touchedSpot != null) {
+                  final sectionIndex = touchResponse.touchedSpot.spotIndex;
+                  // Tap happened
+                  setState(() {
+                    if (selectedSpots.contains(sectionIndex)) {
+                      selectedSpots.remove(sectionIndex);
+                    } else {
+                      selectedSpots.add(sectionIndex);
+                    }
+                  });
                 }
               },
             ),
