@@ -4,10 +4,11 @@ export 'src/chart/bar_chart/bar_chart.dart';
 export 'src/chart/bar_chart/bar_chart_data.dart';
 export 'src/chart/base/axis_chart/axis_chart_data.dart';
 export 'src/chart/base/base_chart/base_chart_data.dart';
-export 'src/chart/base/base_chart/touch_input.dart';
 export 'src/chart/line_chart/line_chart.dart';
 export 'src/chart/line_chart/line_chart_data.dart';
 export 'src/chart/pie_chart/pie_chart.dart';
 export 'src/chart/pie_chart/pie_chart_data.dart';
+export 'src/chart/radar_chart/radar_chart.dart';
+export 'src/chart/radar_chart/radar_chart_data.dart';
 export 'src/chart/scatter_chart/scatter_chart.dart';
 export 'src/chart/scatter_chart/scatter_chart_data.dart';
