@@ -1,6 +1,7 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+
 import '../data_pool.dart';
 
 void main() {
@@ -47,7 +48,7 @@ void main() {
 
       final changed = BackgroundBarChartRodData(
         y: 21,
-        color: Colors.blue,
+        colors: [Colors.blue],
         show: false,
       );
 
@@ -55,7 +56,7 @@ void main() {
 
       final changed2 = BackgroundBarChartRodData(
         y: 22,
-        color: Colors.blue,
+        colors: [Colors.blue],
         show: true,
       );
 
@@ -92,12 +93,6 @@ void main() {
       expect(barTooltipItem1 == barTooltipItem3, false);
       expect(barTooltipItem1 == barTooltipItem4, false);
       expect(barTooltipItem1 == barTooltipItem5, false);
-    });
-
-    test('BarTouchResponse equality test', () {
-      expect(barTouchResponse1 == barTouchResponse1Clone, true);
-      expect(barTouchResponse1 == barTouchResponse2, false);
-      expect(barTouchResponse1 == barTouchResponse3, false);
     });
 
     test('BarTouchedSpot equality test', () {
