@@ -37,8 +37,8 @@ final RangeAnnotations rangeAnnotations2 = RangeAnnotations(horizontalRangeAnnot
 final FlLine flLine1 = FlLine(color: Colors.green, strokeWidth: 1, dashArray: [1, 2, 3]);
 final FlLine flLine1Clone = FlLine(color: Colors.green, strokeWidth: 1, dashArray: [1, 2, 3]);
 
-final Function(double) checkToShowLine = (value) => true;
-final Function(double) getDrawingLine = (value) => FlLine();
+final bool Function(double) checkToShowLine = (value) => true;
+final FlLine Function(double) getDrawingLine = (value) => FlLine();
 
 final FlSpot flSpot1 = FlSpot(1, 1);
 final FlSpot flSpot1Clone = flSpot1.copyWith();
@@ -46,12 +46,14 @@ final FlSpot flSpot1Clone = flSpot1.copyWith();
 final FlSpot flSpot2 = FlSpot(4, 2);
 final FlSpot flSpot2Clone = flSpot2.copyWith();
 
-final Function(double value) getTitles = (value) => 'sallam';
+final String Function(double value) getTitles = (value) => 'sallam';
+final TextStyle Function(double value) getTextStyles =
+    (value) => const TextStyle(color: Colors.green);
 
 final SideTitles sideTitles1 = SideTitles(
   margin: 1,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
@@ -60,7 +62,7 @@ final SideTitles sideTitles1 = SideTitles(
 final SideTitles sideTitles1Clone = SideTitles(
   margin: 1,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
@@ -69,7 +71,7 @@ final SideTitles sideTitles1Clone = SideTitles(
 final SideTitles sideTitles2 = SideTitles(
   margin: 1,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: null,
   interval: 12,
@@ -78,7 +80,7 @@ final SideTitles sideTitles2 = SideTitles(
 final SideTitles sideTitles3 = SideTitles(
   margin: 4,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
@@ -87,7 +89,7 @@ final SideTitles sideTitles3 = SideTitles(
 final SideTitles sideTitles4 = SideTitles(
   margin: 1,
   reservedSize: 11,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
@@ -96,16 +98,16 @@ final SideTitles sideTitles4 = SideTitles(
 final SideTitles sideTitles5 = SideTitles(
   margin: 1,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.red),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
-  rotateAngle: 11,
+  rotateAngle: 101110,
 );
 final SideTitles sideTitles6 = SideTitles(
   margin: 1,
   reservedSize: 10,
-  textStyle: TextStyle(color: Colors.green),
+  getTextStyles: getTextStyles,
   showTitles: false,
   getTitles: getTitles,
   interval: 12,
@@ -163,7 +165,7 @@ final FlTitlesData flTitlesData6 = FlTitlesData(
 );
 
 final AxisTitle axisTitle1 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 33,
   showTitle: true,
@@ -171,7 +173,7 @@ final AxisTitle axisTitle1 = AxisTitle(
   textAlign: TextAlign.right,
 );
 final AxisTitle axisTitle1Clone = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 33,
   showTitle: true,
@@ -180,7 +182,7 @@ final AxisTitle axisTitle1Clone = AxisTitle(
 );
 
 final AxisTitle axisTitle2 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 33,
   margin: 33,
   showTitle: true,
@@ -188,7 +190,7 @@ final AxisTitle axisTitle2 = AxisTitle(
   textAlign: TextAlign.right,
 );
 final AxisTitle axisTitle3 = AxisTitle(
-  textStyle: TextStyle(color: Colors.red),
+  textStyle: const TextStyle(color: Colors.red),
   reservedSize: 12,
   margin: 33,
   showTitle: true,
@@ -196,7 +198,7 @@ final AxisTitle axisTitle3 = AxisTitle(
   textAlign: TextAlign.right,
 );
 final AxisTitle axisTitle4 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 11,
   showTitle: true,
@@ -204,7 +206,7 @@ final AxisTitle axisTitle4 = AxisTitle(
   textAlign: TextAlign.right,
 );
 final AxisTitle axisTitle5 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 33,
   showTitle: false,
@@ -212,7 +214,7 @@ final AxisTitle axisTitle5 = AxisTitle(
   textAlign: TextAlign.right,
 );
 final AxisTitle axisTitle6 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 33,
   showTitle: true,
@@ -220,7 +222,7 @@ final AxisTitle axisTitle6 = AxisTitle(
   textAlign: TextAlign.left,
 );
 final AxisTitle axisTitle7 = AxisTitle(
-  textStyle: TextStyle(color: Colors.green),
+  textStyle: const TextStyle(color: Colors.green),
   reservedSize: 12,
   margin: 33,
   showTitle: true,
@@ -367,12 +369,8 @@ final FlGridData flGridData5 = FlGridData(
   getDrawingVerticalLine: null,
 );
 
-final BaseTouchResponse baseTouchResponse1 = BaseTouchResponse(
-  FlPanStart(Offset.zero),
-);
-final BaseTouchResponse baseTouchResponse2 = BaseTouchResponse(
-  FlPanStart(Offset.zero),
-);
+final BaseTouchResponse baseTouchResponse1 = BaseTouchResponse(new PointerDownEvent(), false);
+final BaseTouchResponse baseTouchResponse2 = BaseTouchResponse(new PointerUpEvent(), true);
 
 final FlTouchData touchData1 = FlTouchData(
   false,
@@ -394,7 +392,7 @@ final FlBorderData borderData2 = FlBorderData(
   border: Border.all(color: Colors.green.withOpacity(0.5)),
 );
 
-final Function(FlSpot) checkToShowSpotLine = (spot) => true;
+final bool Function(FlSpot) checkToShowSpotLine = (spot) => true;
 
 final BarAreaSpotsLine barAreaSpotsLine1 =
     BarAreaSpotsLine(show: true, checkToShowSpotLine: checkToShowSpotLine);
@@ -458,84 +456,65 @@ final BarAreaData barAreaData4 = BarAreaData(
   spotsLine: barAreaSpotsLine2,
 );
 
-final Function(FlSpot, double, LineChartBarData) getDotColor = (spot, percent, bar) => Colors.green;
-final Function(FlSpot spot, LineChartBarData barData) checkToShowDot = (spot, barData) => true;
+final bool Function(FlSpot spot, LineChartBarData barData) checkToShowDot = (spot, barData) => true;
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawer =
+    (spot, percent, barData, index) =>
+        FlDotCirclePainter(radius: 44, color: Colors.green, strokeWidth: 12);
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawer5 =
+    (spot, percent, barData, index) =>
+        FlDotCirclePainter(radius: 44, color: Colors.green, strokeWidth: 14);
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawer6 =
+    (spot, percent, barData, index) =>
+        FlDotCirclePainter(radius: 44.01, color: Colors.green, strokeWidth: 14);
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawerTouched =
+    (spot, percent, barData, index) => FlDotCirclePainter(radius: 12, color: Colors.red);
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawerTouched4 =
+    (spot, percent, barData, index) => FlDotCirclePainter(radius: 12, color: Colors.green);
+final FlDotCirclePainter Function(FlSpot, double, LineChartBarData, int) getDotDrawerTouched6 =
+    (spot, percent, barData, index) => FlDotCirclePainter(radius: 12.01, color: Colors.red);
 
 final FlDotData flDotData1 = FlDotData(
   show: true,
-  strokeWidth: 12,
-  getDotColor: getDotColor,
-  dotSize: 44,
-  getStrokeColor: getDotColor,
+  getDotPainter: getDotDrawer,
   checkToShowDot: checkToShowDot,
 );
 final FlDotData flDotData1Clone = FlDotData(
   show: true,
-  strokeWidth: 12,
-  getDotColor: getDotColor,
-  dotSize: 44,
-  getStrokeColor: getDotColor,
-  checkToShowDot: checkToShowDot,
-);
-
-final FlDotData flDotData2 = FlDotData(
-  show: true,
-  strokeWidth: 12,
-  getDotColor: null,
-  dotSize: 44,
-  getStrokeColor: getDotColor,
-  checkToShowDot: checkToShowDot,
-);
-
-final FlDotData flDotData3 = FlDotData(
-  show: true,
-  strokeWidth: 12,
-  getDotColor: getDotColor,
-  dotSize: 44,
-  getStrokeColor: null,
+  getDotPainter: getDotDrawer,
   checkToShowDot: checkToShowDot,
 );
 
 final FlDotData flDotData4 = FlDotData(
   show: true,
-  strokeWidth: 12,
-  getDotColor: getDotColor,
-  dotSize: 44,
-  getStrokeColor: getDotColor,
+  getDotPainter: getDotDrawer,
   checkToShowDot: null,
 );
 
 final FlDotData flDotData5 = FlDotData(
   show: true,
-  strokeWidth: 14,
-  getDotColor: getDotColor,
-  dotSize: 44,
-  getStrokeColor: getDotColor,
+  getDotPainter: getDotDrawer5,
   checkToShowDot: null,
 );
 
 final FlDotData flDotData6 = FlDotData(
   show: true,
-  strokeWidth: 14,
-  getDotColor: getDotColor,
-  dotSize: 44.01,
-  getStrokeColor: getDotColor,
+  getDotPainter: getDotDrawer6,
   checkToShowDot: null,
 );
 
-final Shadow shadow1 = Shadow(
+const Shadow shadow1 = Shadow(
   color: Colors.red,
   blurRadius: 12,
 );
-final Shadow shadow1Clone = Shadow(
+const Shadow shadow1Clone = Shadow(
   color: Colors.red,
   blurRadius: 12,
 );
-final Shadow shadow2 = Shadow(
+const Shadow shadow2 = Shadow(
   color: Colors.green,
   blurRadius: 12,
 );
-final Shadow shadow3 = Shadow(
+const Shadow shadow3 = Shadow(
   color: Colors.red,
   blurRadius: 14,
 );
@@ -721,7 +700,7 @@ final LineChartBarData lineChartBarData6 = LineChartBarData(
   gradientTo: const Offset(1, 1),
   barWidth: 12,
   curveSmoothness: 12.0,
-  dotData: flDotData2,
+  dotData: flDotData1,
   isCurved: false,
   isStrokeCapRound: true,
   preventCurveOverShooting: false,
@@ -826,14 +805,16 @@ final LineTouchResponse lineTouchResponse1 = LineTouchResponse(
     lineBarSpot1,
     lineBarSpot2,
   ],
-  FlPanStart(Offset.zero),
+  new PointerDownEvent(),
+  false,
 );
 final LineTouchResponse lineTouchResponse1Clone = LineTouchResponse(
   [
     lineBarSpot1Clone,
     lineBarSpot2,
   ],
-  FlPanStart(Offset.zero),
+  new PointerDownEvent(),
+  false,
 );
 
 final LineTouchResponse lineTouchResponse2 = LineTouchResponse(
@@ -841,12 +822,14 @@ final LineTouchResponse lineTouchResponse2 = LineTouchResponse(
     lineBarSpot2,
     lineBarSpot1,
   ],
-  FlPanStart(Offset.zero),
+  new PointerDownEvent(),
+  false,
 );
 
 final LineTouchResponse lineTouchResponse3 = LineTouchResponse(
-  null,
-  FlPanStart(Offset.zero),
+  [],
+  new PointerDownEvent(position: Offset.zero),
+  false,
 );
 
 final LineTouchResponse lineTouchResponse4 = LineTouchResponse(
@@ -854,7 +837,8 @@ final LineTouchResponse lineTouchResponse4 = LineTouchResponse(
     lineBarSpot1,
     lineBarSpot2,
   ],
-  null,
+  new PointerMoveEvent(position: const Offset(1, 2), delta: Offset(1, 1)),
+  false,
 );
 
 final LineTouchResponse lineTouchResponse5 = LineTouchResponse(
@@ -862,7 +846,8 @@ final LineTouchResponse lineTouchResponse5 = LineTouchResponse(
     lineBarSpot1,
     lineBarSpot2,
   ],
-  FlPanStart(const Offset(0, 100)),
+  new PointerDownEvent(position: const Offset(0, 100)),
+  false,
 );
 
 final TouchedSpotIndicatorData touchedSpotIndicatorData1 = TouchedSpotIndicatorData(
@@ -870,14 +855,14 @@ final TouchedSpotIndicatorData touchedSpotIndicatorData1 = TouchedSpotIndicatorD
     color: Colors.red,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12, checkToShowDot: checkToShowDot, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched, checkToShowDot: checkToShowDot, show: true),
 );
 final TouchedSpotIndicatorData touchedSpotIndicatorData1Clone = TouchedSpotIndicatorData(
   FlLine(
     color: Colors.red,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12, checkToShowDot: checkToShowDot, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched, checkToShowDot: checkToShowDot, show: true),
 );
 
 final TouchedSpotIndicatorData touchedSpotIndicatorData2 = TouchedSpotIndicatorData(
@@ -885,45 +870,48 @@ final TouchedSpotIndicatorData touchedSpotIndicatorData2 = TouchedSpotIndicatorD
     color: Colors.red,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12, checkToShowDot: null, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched, checkToShowDot: null, show: true),
 );
 final TouchedSpotIndicatorData touchedSpotIndicatorData3 = TouchedSpotIndicatorData(
   FlLine(
     color: Colors.red,
     dashArray: null,
   ),
-  FlDotData(dotSize: 12, checkToShowDot: checkToShowDot, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched, checkToShowDot: checkToShowDot, show: true),
 );
 final TouchedSpotIndicatorData touchedSpotIndicatorData4 = TouchedSpotIndicatorData(
   FlLine(
     color: Colors.green,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12, checkToShowDot: checkToShowDot, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched4, checkToShowDot: checkToShowDot, show: true),
 );
 final TouchedSpotIndicatorData touchedSpotIndicatorData5 = TouchedSpotIndicatorData(
   FlLine(
     color: Colors.red,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12, checkToShowDot: checkToShowDot, show: false),
+  FlDotData(getDotPainter: getDotDrawerTouched, checkToShowDot: checkToShowDot, show: false),
 );
 final TouchedSpotIndicatorData touchedSpotIndicatorData6 = TouchedSpotIndicatorData(
   FlLine(
     color: Colors.red,
     dashArray: [],
   ),
-  FlDotData(dotSize: 12.01, checkToShowDot: checkToShowDot, show: true),
+  FlDotData(getDotPainter: getDotDrawerTouched6, checkToShowDot: checkToShowDot, show: true),
 );
 
-final LineTooltipItem lineTooltipItem1 = LineTooltipItem('', TextStyle(color: Colors.green));
-final LineTooltipItem lineTooltipItem1Clone = LineTooltipItem('', TextStyle(color: Colors.green));
+final LineTooltipItem lineTooltipItem1 = LineTooltipItem('', const TextStyle(color: Colors.green));
+final LineTooltipItem lineTooltipItem1Clone =
+    LineTooltipItem('', const TextStyle(color: Colors.green));
 
-final LineTooltipItem lineTooltipItem2 = LineTooltipItem('ss', TextStyle(color: Colors.green));
-final LineTooltipItem lineTooltipItem3 = LineTooltipItem('', TextStyle(color: Colors.blue));
-final LineTooltipItem lineTooltipItem4 = LineTooltipItem('', null);
+final LineTooltipItem lineTooltipItem2 =
+    LineTooltipItem('ss', const TextStyle(color: Colors.green));
+final LineTooltipItem lineTooltipItem3 = LineTooltipItem('', const TextStyle(color: Colors.blue));
+final LineTooltipItem lineTooltipItem4 = LineTooltipItem('', const TextStyle(fontSize: 33));
 
-final Function(List<LineBarSpot> touchedSpots) lineChartGetTooltipItems = (list) {
+final List<LineTooltipItem?> Function(List<LineBarSpot> touchedSpots) lineChartGetTooltipItems =
+    (list) {
   return list.map((s) => lineTooltipItem1).toList();
 };
 
@@ -935,7 +923,7 @@ final LineTouchTooltipData lineTouchTooltipData1 = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 33,
+  tooltipMargin: 33,
 );
 final LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
   tooltipPadding: const EdgeInsets.all(0.1),
@@ -945,7 +933,7 @@ final LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 33,
+  tooltipMargin: 33,
 );
 
 final LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
@@ -956,7 +944,7 @@ final LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 33,
+  tooltipMargin: 33,
 );
 final LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
   tooltipPadding: const EdgeInsets.all(0.2),
@@ -966,7 +954,7 @@ final LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 33,
+  tooltipMargin: 33,
 );
 final LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
   tooltipPadding: const EdgeInsets.all(0.1),
@@ -976,7 +964,7 @@ final LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 33,
+  tooltipMargin: 33,
 );
 final LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
   tooltipPadding: const EdgeInsets.all(0.1),
@@ -986,12 +974,13 @@ final LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
   fitInsideHorizontally: true,
   fitInsideVertically: false,
   tooltipRoundedRadius: 12,
-  tooltipBottomMargin: 34,
+  tooltipMargin: 34,
 );
 
 final Function(LineTouchResponse) lineTouchCallback = (response) {};
 
-final Function(LineChartBarData barData, List<int> spotIndexes) getTouchedSpotIndicator =
+final List<TouchedSpotIndicatorData?> Function(LineChartBarData barData, List<int> spotIndexes)
+    getTouchedSpotIndicator =
     (barData, indexes) => indexes.map((i) => touchedSpotIndicatorData1).toList();
 
 final LineTouchData lineTouchData1 = LineTouchData(
@@ -1068,26 +1057,26 @@ final LineTouchData lineTouchData7 = LineTouchData(
   fullHeightTouchLine: true,
 );
 
-final Function(HorizontalLine) horizontalLabelResolver = (horizontalLine) => 'test';
-final Function(VerticalLine) verticalLabelResolver = (horizontalLine) => 'test';
+final String Function(HorizontalLine) horizontalLabelResolver = (horizontalLine) => 'test';
+final String Function(VerticalLine) verticalLabelResolver = (horizontalLine) => 'test';
 
 final HorizontalLineLabel horizontalLineLabel1 = HorizontalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final HorizontalLineLabel horizontalLineLabel1Clone = HorizontalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final HorizontalLineLabel horizontalLineLabel2 = HorizontalLineLabel(
   show: false,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
@@ -1101,28 +1090,28 @@ final HorizontalLineLabel horizontalLineLabel3 = HorizontalLineLabel(
 );
 final HorizontalLineLabel horizontalLineLabel4 = HorizontalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: null,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final HorizontalLineLabel horizontalLineLabel5 = HorizontalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.bottomRight,
   padding: const EdgeInsets.all(12),
 );
 final HorizontalLineLabel horizontalLineLabel6 = HorizontalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(44),
 );
 final HorizontalLineLabel horizontalLineLabel7 = HorizontalLineLabel(
   show: false,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: horizontalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
@@ -1130,21 +1119,21 @@ final HorizontalLineLabel horizontalLineLabel7 = HorizontalLineLabel(
 
 final VerticalLineLabel verticalLineLabel1 = VerticalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final VerticalLineLabel verticalLineLabel1Clone = VerticalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final VerticalLineLabel verticalLineLabel2 = VerticalLineLabel(
   show: false,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
@@ -1158,28 +1147,28 @@ final VerticalLineLabel verticalLineLabel3 = VerticalLineLabel(
 );
 final VerticalLineLabel verticalLineLabel4 = VerticalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: null,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
 );
 final VerticalLineLabel verticalLineLabel5 = VerticalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.bottomRight,
   padding: const EdgeInsets.all(12),
 );
 final VerticalLineLabel verticalLineLabel6 = VerticalLineLabel(
   show: true,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(44),
 );
 final VerticalLineLabel verticalLineLabel7 = VerticalLineLabel(
   show: false,
-  style: TextStyle(color: Colors.green),
+  style: const TextStyle(color: Colors.green),
   labelResolver: verticalLabelResolver,
   alignment: Alignment.topCenter,
   padding: const EdgeInsets.all(12),
@@ -1192,7 +1181,7 @@ final HorizontalLine horizontalLine1 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine1Clone = HorizontalLine(
   y: 12,
@@ -1201,12 +1190,12 @@ final HorizontalLine horizontalLine1Clone = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine2 = HorizontalLine(
   y: 12,
   color: Colors.red,
-  dashArray: [0, 1],
+  dashArray: [0, 1, 3],
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
@@ -1219,7 +1208,7 @@ final HorizontalLine horizontalLine3 = HorizontalLine(
   strokeWidth: 22,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine4 = HorizontalLine(
   y: 12,
@@ -1228,7 +1217,7 @@ final HorizontalLine horizontalLine4 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine5 = HorizontalLine(
   y: 33,
@@ -1237,7 +1226,7 @@ final HorizontalLine horizontalLine5 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine6 = HorizontalLine(
   y: 12,
@@ -1246,7 +1235,7 @@ final HorizontalLine horizontalLine6 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine7 = HorizontalLine(
   y: 12,
@@ -1255,7 +1244,7 @@ final HorizontalLine horizontalLine7 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel2,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine8 = HorizontalLine(
   y: 12,
@@ -1264,16 +1253,16 @@ final HorizontalLine horizontalLine8 = HorizontalLine(
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final HorizontalLine horizontalLine9 = HorizontalLine(
   y: 12,
   color: Colors.red,
-  dashArray: [0, 1],
+  dashArray: [0, 12, 44],
   strokeWidth: 21,
   label: horizontalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 15, 15),
+  sizedPicture: null,
 );
 
 final VerticalLine verticalLine1 = VerticalLine(
@@ -1283,7 +1272,7 @@ final VerticalLine verticalLine1 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine1Clone = VerticalLine(
   x: 12,
@@ -1292,11 +1281,11 @@ final VerticalLine verticalLine1Clone = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine2 = VerticalLine(
   x: 12,
-  color: Colors.red,
+  color: Colors.green,
   dashArray: [0, 1],
   strokeWidth: 21,
   label: verticalLineLabel1,
@@ -1310,7 +1299,7 @@ final VerticalLine verticalLine3 = VerticalLine(
   strokeWidth: 22,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine4 = VerticalLine(
   x: 12,
@@ -1319,7 +1308,7 @@ final VerticalLine verticalLine4 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine5 = VerticalLine(
   x: 33,
@@ -1328,7 +1317,7 @@ final VerticalLine verticalLine5 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine6 = VerticalLine(
   x: 12,
@@ -1337,7 +1326,7 @@ final VerticalLine verticalLine6 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine7 = VerticalLine(
   x: 12,
@@ -1346,7 +1335,7 @@ final VerticalLine verticalLine7 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel2,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine8 = VerticalLine(
   x: 12,
@@ -1355,16 +1344,16 @@ final VerticalLine verticalLine8 = VerticalLine(
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 14, 14),
+  sizedPicture: null,
 );
 final VerticalLine verticalLine9 = VerticalLine(
   x: 12,
   color: Colors.red,
-  dashArray: [0, 1],
+  dashArray: [0, 12, 44],
   strokeWidth: 21,
   label: verticalLineLabel1,
   image: null,
-  sizedPicture: SizedPicture(null, 15, 15),
+  sizedPicture: null,
 );
 
 final ExtraLinesData extraLinesData1 = ExtraLinesData(
@@ -1473,7 +1462,7 @@ final SizedPicture sizedPicture3 = SizedPicture(
 );
 final SizedPicture sizedPicture4 = SizedPicture(
   PictureRecorder().endRecording(),
-  null,
+  442,
   30,
 );
 
@@ -1559,8 +1548,8 @@ final ShowingTooltipIndicators showingTooltipIndicator1Clone = ShowingTooltipInd
   [lineBarSpot1, lineBarSpot2],
 );
 final ShowingTooltipIndicators showingTooltipIndicator2 = ShowingTooltipIndicators(
-  1,
-  null,
+  33,
+  [],
 );
 final ShowingTooltipIndicators showingTooltipIndicator3 = ShowingTooltipIndicators(
   1,
@@ -1583,7 +1572,7 @@ final LineChartData lineChartData1 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1604,7 +1593,7 @@ final LineChartData lineChartData1Clone = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1Clone,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1625,7 +1614,7 @@ final LineChartData lineChartData2 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1646,7 +1635,7 @@ final LineChartData lineChartData3 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1664,7 +1653,7 @@ final LineChartData lineChartData4 = LineChartData(
   lineTouchData: lineTouchData1,
   showingTooltipIndicators: [],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1685,7 +1674,7 @@ final LineChartData lineChartData5 = LineChartData(
     showingTooltipIndicator1,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1703,7 +1692,7 @@ final LineChartData lineChartData6 = LineChartData(
   lineTouchData: lineTouchData1,
   showingTooltipIndicators: null,
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1724,7 +1713,7 @@ final LineChartData lineChartData7 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData2,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1745,7 +1734,7 @@ final LineChartData lineChartData8 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: true,
+  clipData: FlClipData.all(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1766,7 +1755,7 @@ final LineChartData lineChartData9 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red.withOpacity(0.2),
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1787,7 +1776,7 @@ final LineChartData lineChartData10 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 24,
   rangeAnnotations: rangeAnnotations1,
@@ -1808,7 +1797,7 @@ final LineChartData lineChartData11 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: null,
@@ -1829,7 +1818,7 @@ final LineChartData lineChartData12 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1850,7 +1839,7 @@ final LineChartData lineChartData13 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1871,7 +1860,7 @@ final LineChartData lineChartData14 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1892,7 +1881,7 @@ final LineChartData lineChartData15 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1913,7 +1902,7 @@ final LineChartData lineChartData16 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1934,7 +1923,7 @@ final LineChartData lineChartData17 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1955,7 +1944,7 @@ final LineChartData lineChartData18 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1976,7 +1965,7 @@ final LineChartData lineChartData19 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -1997,7 +1986,7 @@ final LineChartData lineChartData20 = LineChartData(
     showingTooltipIndicator2,
   ],
   axisTitleData: flAxisTitleData1,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   backgroundColor: Colors.red,
   maxY: 23,
   rangeAnnotations: rangeAnnotations1,
@@ -2027,12 +2016,19 @@ final PieChartData pieChartData1 = PieChartData(
 );
 final PieChartData pieChartData1Clone = pieChartData1.copyWith();
 
-final Function(double) gridCheckToShowLine = (value) => true;
-final Function(double) gridGetDrawingLine = (value) => FlLine();
+final bool Function(double) gridCheckToShowLine = (value) => true;
+final FlLine Function(double) gridGetDrawingLine = (value) => FlLine();
 
-final Function(ScatterSpot touchedSpots) scatterChartGetTooltipItems = (list) {
-  return ScatterTooltipItem('check', TextStyle(color: Colors.blue), 23);
+final ScatterTooltipItem? Function(ScatterSpot touchedSpots) scatterChartGetTooltipItems = (list) {
+  return ScatterTooltipItem('check', const TextStyle(color: Colors.blue), 23);
 };
+
+final ScatterSpot scatterSpot1 = ScatterSpot(1, 40);
+final ScatterSpot scatterSpot1Clone = ScatterSpot(1, 40);
+final ScatterSpot scatterSpot2 = ScatterSpot(-4, -8);
+final ScatterSpot scatterSpot2Clone = scatterSpot2.copyWith();
+final ScatterSpot scatterSpot3 = ScatterSpot(-14, 5);
+final ScatterSpot scatterSpot4 = ScatterSpot(-0, 0);
 
 final ScatterChartData scatterChartData1 = ScatterChartData(
   minY: 0,
@@ -2043,7 +2039,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
     show: true,
     leftTitle: AxisTitle(
       showTitle: true,
-      textStyle: TextStyle(color: Colors.red, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.red, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 22,
       margin: 11,
@@ -2051,7 +2047,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
     ),
     bottomTitle: AxisTitle(
       showTitle: false,
-      textStyle: TextStyle(color: Colors.grey, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.grey, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 11,
       margin: 11,
@@ -2059,7 +2055,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
     ),
     rightTitle: AxisTitle(
       showTitle: false,
-      textStyle: TextStyle(color: Colors.blue, fontSize: 11),
+      textStyle: const TextStyle(color: Colors.blue, fontSize: 11),
       textAlign: TextAlign.left,
       reservedSize: 2,
       margin: 1324,
@@ -2067,7 +2063,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
     ),
     topTitle: AxisTitle(
       showTitle: true,
-      textStyle: TextStyle(color: Colors.green, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.green, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 23,
       margin: 11,
@@ -2086,7 +2082,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
     verticalInterval: 1,
   ),
   backgroundColor: Colors.black,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   borderData: FlBorderData(
       show: true,
       border: Border.all(
@@ -2130,7 +2126,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
     show: true,
     leftTitle: AxisTitle(
       showTitle: true,
-      textStyle: TextStyle(color: Colors.red, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.red, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 22,
       margin: 11,
@@ -2138,7 +2134,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
     ),
     bottomTitle: AxisTitle(
       showTitle: false,
-      textStyle: TextStyle(color: Colors.grey, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.grey, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 11,
       margin: 11,
@@ -2146,7 +2142,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
     ),
     rightTitle: AxisTitle(
       showTitle: false,
-      textStyle: TextStyle(color: Colors.blue, fontSize: 11),
+      textStyle: const TextStyle(color: Colors.blue, fontSize: 11),
       textAlign: TextAlign.left,
       reservedSize: 2,
       margin: 1324,
@@ -2154,7 +2150,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
     ),
     topTitle: AxisTitle(
       showTitle: true,
-      textStyle: TextStyle(color: Colors.green, fontSize: 33),
+      textStyle: const TextStyle(color: Colors.green, fontSize: 33),
       textAlign: TextAlign.left,
       reservedSize: 23,
       margin: 11,
@@ -2173,7 +2169,7 @@ final ScatterChartData scatterChartData1Clone = ScatterChartData(
     verticalInterval: 1,
   ),
   backgroundColor: Colors.black,
-  clipToBorder: false,
+  clipData: FlClipData.none(),
   borderData: FlBorderData(
       show: true,
       border: Border.all(
@@ -2224,31 +2220,31 @@ final BarChartRodStackItem barChartRodStackItem2 = BarChartRodStackItem(
 
 final BackgroundBarChartRodData backgroundBarChartRodData1 = BackgroundBarChartRodData(
   y: 21,
-  color: Colors.blue,
+  colors: [Colors.blue],
   show: true,
 );
 final BackgroundBarChartRodData backgroundBarChartRodData1Clone = BackgroundBarChartRodData(
   y: 21,
-  color: Colors.blue,
+  colors: [Colors.blue],
   show: true,
 );
 final BackgroundBarChartRodData backgroundBarChartRodData2 = BackgroundBarChartRodData(
   y: 44,
-  color: Colors.red,
+  colors: [Colors.red],
   show: true,
 );
 final BackgroundBarChartRodData backgroundBarChartRodData3 = BackgroundBarChartRodData(
   y: 44,
-  color: Colors.green,
+  colors: [Colors.green],
   show: true,
 );
 
 final BarChartRodData barChartRodData1 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1,
     barChartRodStackItem2,
   ],
@@ -2256,80 +2252,80 @@ final BarChartRodData barChartRodData1 = BarChartRodData(
 );
 
 final BarChartRodData barChartRodData1Clone = barChartRodData1.copyWith(
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1Clone,
     barChartRodStackItem2,
   ],
 );
 
 final BarChartRodData barChartRodData2 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 1132,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1,
     barChartRodStackItem2,
   ],
   backDrawRodData: backgroundBarChartRodData1,
 );
 final BarChartRodData barChartRodData3 = BarChartRodData(
-  color: Colors.green,
+  colors: [Colors.green],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem2,
   ],
   backDrawRodData: backgroundBarChartRodData1,
 );
 final BarChartRodData barChartRodData4 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem2,
     barChartRodStackItem1,
   ],
   backDrawRodData: backgroundBarChartRodData1,
 );
 final BarChartRodData barChartRodData5 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 55,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1,
     barChartRodStackItem2,
   ],
   backDrawRodData: backgroundBarChartRodData1,
 );
 final BarChartRodData barChartRodData6 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: null,
+  rodStackItems: null,
   backDrawRodData: backgroundBarChartRodData1,
 );
 final BarChartRodData barChartRodData7 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(12)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1,
     barChartRodStackItem2,
   ],
   backDrawRodData: backgroundBarChartRodData2,
 );
 final BarChartRodData barChartRodData8 = BarChartRodData(
-  color: Colors.red,
+  colors: [Colors.red],
   y: 12,
   width: 32,
   borderRadius: const BorderRadius.all(Radius.circular(14)),
-  rodStackItem: [
+  rodStackItems: [
     barChartRodStackItem1,
     barChartRodStackItem2,
   ],
@@ -2446,6 +2442,8 @@ final BarTouchedSpot barTouchedSpot1 = BarTouchedSpot(
   1,
   barChartRodData1,
   2,
+  barChartRodStackItem1,
+  1,
   flSpot1,
   Offset.zero,
 );
@@ -2454,6 +2452,8 @@ final BarTouchedSpot barTouchedSpot1Clone = BarTouchedSpot(
   1,
   barChartRodData1Clone,
   2,
+  barChartRodStackItem1Clone,
+  1,
   flSpot1Clone,
   Offset.zero,
 );
@@ -2461,6 +2461,8 @@ final BarTouchedSpot barTouchedSpot2 = BarTouchedSpot(
   barChartGroupData2,
   1,
   barChartRodData1,
+  2,
+  barChartRodStackItem2,
   2,
   flSpot1,
   Offset.zero,
@@ -2470,6 +2472,8 @@ final BarTouchedSpot barTouchedSpot3 = BarTouchedSpot(
   1,
   barChartRodData2,
   2,
+  barChartRodStackItem2,
+  2,
   flSpot1,
   Offset.zero,
 );
@@ -2477,6 +2481,8 @@ final BarTouchedSpot barTouchedSpot4 = BarTouchedSpot(
   barChartGroupData1,
   2,
   barChartRodData1,
+  2,
+  barChartRodStackItem2,
   2,
   flSpot1,
   Offset.zero,
@@ -2486,6 +2492,8 @@ final BarTouchedSpot barTouchedSpot5 = BarTouchedSpot(
   1,
   barChartRodData1,
   3,
+  barChartRodStackItem2,
+  2,
   flSpot1,
   Offset.zero,
 );
@@ -2493,6 +2501,8 @@ final BarTouchedSpot barTouchedSpot6 = BarTouchedSpot(
   barChartGroupData1,
   1,
   barChartRodData1,
+  2,
+  barChartRodStackItem2,
   2,
   flSpot2,
   Offset.zero,
@@ -2502,50 +2512,51 @@ final BarTouchedSpot barTouchedSpot7 = BarTouchedSpot(
   1,
   barChartRodData1,
   2,
+  barChartRodStackItem2,
+  2,
   flSpot1,
   const Offset(1, 10),
 );
 
 final BarTouchResponse barTouchResponse1 = BarTouchResponse(
   barTouchedSpot1,
-  FlPanStart(const Offset(0, 1)),
+  new PointerDownEvent(position: const Offset(0, 1)),
+  false,
 );
 final BarTouchResponse barTouchResponse1Clone = BarTouchResponse(
   barTouchedSpot1Clone,
-  FlPanStart(const Offset(0, 1)),
+  new PointerDownEvent(position: const Offset(0, 1)),
+  false,
 );
 final BarTouchResponse barTouchResponse2 = BarTouchResponse(
   barTouchedSpot2,
-  FlPanStart(const Offset(0, 1)),
-);
-final BarTouchResponse barTouchResponse3 = BarTouchResponse(
-  barTouchedSpot1,
-  FlPanStart(const Offset(0.1, 1)),
+  new PointerDownEvent(position: const Offset(0, 1)),
+  false,
 );
 
 final BarTooltipItem barTooltipItem1 = BarTooltipItem(
   'pashmam 1',
-  TextStyle(color: Colors.red),
+  const TextStyle(color: Colors.red),
 );
 final BarTooltipItem barTooltipItem1Clone = BarTooltipItem(
   'pashmam 1',
-  TextStyle(color: Colors.red),
+  const TextStyle(color: Colors.red),
 );
 final BarTooltipItem barTooltipItem2 = BarTooltipItem(
   'pashmam 2',
-  TextStyle(color: Colors.red),
+  const TextStyle(color: Colors.red),
 );
 final BarTooltipItem barTooltipItem3 = BarTooltipItem(
   'pashmam 1',
-  TextStyle(color: Colors.green),
+  const TextStyle(color: Colors.green),
 );
 final BarTooltipItem barTooltipItem4 = BarTooltipItem(
-  null,
-  TextStyle(color: Colors.red),
+  'null',
+  const TextStyle(color: Colors.red),
 );
 final BarTooltipItem barTooltipItem5 = BarTooltipItem(
   'pashmam 1',
-  null,
+  const TextStyle(fontSize: 85),
 );
 
 BarTooltipItem getTooltipItem(
@@ -2554,7 +2565,7 @@ BarTooltipItem getTooltipItem(
   BarChartRodData rod,
   int rodIndex,
 ) {
-  final TextStyle textStyle = TextStyle(
+  const TextStyle textStyle = TextStyle(
     color: Colors.black,
     fontWeight: FontWeight.bold,
     fontSize: 14,
@@ -2570,7 +2581,7 @@ final BarTouchTooltipData barTouchTooltipData1 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData1Clone = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2580,7 +2591,7 @@ final BarTouchTooltipData barTouchTooltipData1Clone = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData2 = BarTouchTooltipData(
   tooltipRoundedRadius: 13,
@@ -2590,7 +2601,7 @@ final BarTouchTooltipData barTouchTooltipData2 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData3 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2600,7 +2611,7 @@ final BarTouchTooltipData barTouchTooltipData3 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData4 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2610,7 +2621,7 @@ final BarTouchTooltipData barTouchTooltipData4 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData5 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2620,7 +2631,7 @@ final BarTouchTooltipData barTouchTooltipData5 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData6 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2630,7 +2641,7 @@ final BarTouchTooltipData barTouchTooltipData6 = BarTouchTooltipData(
   tooltipBgColor: Colors.blue,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData7 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2640,7 +2651,7 @@ final BarTouchTooltipData barTouchTooltipData7 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: null,
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData8 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2650,7 +2661,7 @@ final BarTouchTooltipData barTouchTooltipData8 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: null,
-  tooltipBottomMargin: 12,
+  tooltipMargin: 12,
 );
 final BarTouchTooltipData barTouchTooltipData9 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
@@ -2660,7 +2671,7 @@ final BarTouchTooltipData barTouchTooltipData9 = BarTouchTooltipData(
   tooltipBgColor: Colors.green,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
-  tooltipBottomMargin: 333,
+  tooltipMargin: 333,
 );
 
 final Function(BarTouchResponse) barTouchCallback = (response) {};
@@ -2814,4 +2825,147 @@ final BarChartData barChartData13 = barChartData1.copyWith(
 );
 final BarChartData barChartData14 = barChartData1.copyWith(
   groupsSpace: 444,
+);
+
+final RadarDataSet radarDataSet1 = RadarDataSet(
+  dataEntries: [
+    RadarEntry(value: 0),
+    RadarEntry(value: 1),
+    RadarEntry(value: 2),
+    RadarEntry(value: 3),
+    RadarEntry(value: 4),
+  ],
+  borderColor: Colors.blue,
+  borderWidth: 3,
+  entryRadius: 3,
+  fillColor: Colors.grey,
+);
+
+final RadarDataSet radarDataSet1Clone = radarDataSet1.copyWith();
+
+final RadarDataSet radarDataSet2 = RadarDataSet(
+  dataEntries: [
+    RadarEntry(value: 10),
+    RadarEntry(value: 9),
+    RadarEntry(value: 8),
+    RadarEntry(value: 7),
+    RadarEntry(value: 6),
+  ],
+  borderColor: Colors.red,
+  borderWidth: 5,
+  entryRadius: 5,
+  fillColor: Colors.black,
+);
+
+final RadarTouchData radarTouchData1 = RadarTouchData(
+  enabled: true,
+  touchCallback: radarTouchCallback,
+  touchSpotThreshold: 12,
+);
+
+final RadarTouchData radarTouchData2 = RadarTouchData(
+  enabled: false,
+  touchCallback: radarTouchCallback,
+  touchSpotThreshold: 5,
+);
+
+final RadarTouchData radarTouchData1Clone = radarTouchData1;
+
+final Function(RadarTouchResponse) radarTouchCallback = (response) {};
+
+final radarTouchedSpot1 = RadarTouchedSpot(
+  radarDataSet1,
+  0,
+  radarDataSet1.dataEntries.first,
+  0,
+  flSpot1,
+  Offset.zero,
+);
+
+final radarTouchedSpotClone1 = radarTouchedSpot1;
+
+final radarTouchedSpot2 = RadarTouchedSpot(
+  radarDataSet2,
+  0,
+  radarDataSet1.dataEntries.first,
+  0,
+  flSpot1,
+  Offset.zero,
+);
+
+final radarTouchedSpot3 = RadarTouchedSpot(
+  radarDataSet1,
+  1,
+  radarDataSet1.dataEntries.first,
+  0,
+  flSpot1,
+  Offset.zero,
+);
+
+final radarTouchedSpot4 = RadarTouchedSpot(
+  radarDataSet1,
+  0,
+  radarDataSet1.dataEntries.last,
+  0,
+  flSpot1,
+  Offset.zero,
+);
+
+final radarTouchedSpot5 = RadarTouchedSpot(
+  radarDataSet1,
+  0,
+  radarDataSet1.dataEntries.first,
+  1,
+  flSpot1,
+  Offset.zero,
+);
+
+final radarTouchedSpot6 = RadarTouchedSpot(
+  radarDataSet1,
+  0,
+  radarDataSet1.dataEntries.first,
+  0,
+  flSpot2,
+  Offset.zero,
+);
+
+final radarTouchedSpot7 = RadarTouchedSpot(
+  radarDataSet1,
+  0,
+  radarDataSet1.dataEntries.first,
+  0,
+  flSpot1,
+  const Offset(3, 5),
+);
+
+final RadarChartData radarChartData1 = RadarChartData(
+  dataSets: [radarDataSet1],
+  radarBackgroundColor: Colors.yellow,
+  radarBorderData: BorderSide(color: Colors.purple, width: 5),
+  borderData: borderData1,
+  gridBorderData: BorderSide(color: Colors.blue, width: 2),
+  getTitle: (index) => 'testTitle',
+  titlePositionPercentageOffset: 0.2,
+  titleTextStyle: TextStyle(color: Colors.white, fontSize: 12),
+  radarTouchData: radarTouchData1,
+  tickCount: 5,
+  tickBorderData: BorderSide(color: Colors.black, width: 4),
+  ticksTextStyle: TextStyle(color: Colors.white, fontSize: 12),
+);
+
+final RadarChartData radarChartData1Clone = radarChartData1.copyWith();
+
+final RadarChartData radarChartData2 = RadarChartData(
+  dataSets: [radarDataSet2],
+  radarBackgroundColor: Colors.blue,
+  radarBorderData: BorderSide(color: Colors.pink, width: 3),
+  borderData: borderData1,
+  gridBorderData: BorderSide(color: Colors.red, width: 3),
+  getTitle: (index) => 'testTitle2',
+  titlePositionPercentageOffset: 0.5,
+  titleTextStyle: TextStyle(color: Colors.black, fontSize: 5),
+  radarTouchData: radarTouchData2,
+  tickCount: 1,
+  tickBorderData: BorderSide(color: Colors.pink, width: 2),
+  ticksTextStyle: TextStyle(color: Colors.purple, fontSize: 10),
 );
